==> scripts/1_2_cleaning_NOAA_EBS_data.Rmd <==
---
title: "1_2_cleaning_NOAA_EBS_data"
author: "Aspen Coyle"
date: "6/26/2021"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

## Background

This script deals with the NOAA EBS data on _Hematodinium_ infection, obtained by the Roberts lab from Pam Jensen. These data contain _Hematodinium_ test infection results from the annual NOAA Eastern Bering Sea (EBS) trawl surveys from 2014 to 2019.

Since files do not have consistent formatting and organization, each is read in separately rather than in a single loop.

#### Load libraries (and install if necessary)

```{r libraries, message=FALSE, warning=FALSE}
# Add all required libraries here
list.of.packages <- c("tidyverse", "readxl", "lubridate", "rnaturalearth", "rnaturalearthdata", "sf", "rgeos")
# Get names of all required packages that aren't installed
new.packages <- list.of.packages[!(list.of.packages %in% installed.packages()[, "Package"])]
# Install all new packages
if(length(new.packages)) install.packages(new.packages)


# Load all required libraries
lapply(list.of.packages, FUN = function(X) {
  do.call("require", list(X))
})
```
#### See all files to be read in
```{r}
# Get a list of all our data files
files <- paste0("../data/NOAA_EBS_trawl_survey/", list.files(path = "../data/NOAA_EBS_trawl_survey/"))
# Display all files we'll be reading in
files
```


#### Read in 2014 data
```{r}
# Read in file from 2014
full_data <- read_excel(files[1], sheet = 2)

# See first few lines of full data sheet
head(full_data)
```

#### Read in 2015 data
```{r}
# Read in file from 2015
new_data <- read_excel(files[2], sheet = 2)

# Check if names match between the two years
all_equal(new_data, full_data, ignore_col_order = FALSE)

# Appears that the differences are solely in formatting (Chela -> chela and STATIONID -> STATION ID). Let's fix those so the 2015 data match the 2014 data
new_data <- new_data %>%
  dplyr::rename(
    Chela = chela,
    STATIONID = "STATION ID"
  )

# See if we have any other mismatches
all_equal(new_data, full_data, ignore_col_order = FALSE, convert = TRUE)

# Join the 2015 data to the 2014 data
full_data <- bind_rows(full_data, new_data)
```

#### Read in 2016 data
```{r}
# Read in file from 2016
new_data <- read_excel(files[3], sheet = 2)

# Check if names match between 2016 data and our full_data table
all_equal(new_data, full_data, ignore_col_order = FALSE)

# Again, appears that the mismatches in column names are simple differences in name formatting, not overall sheet organization. Therefore we can simply rename our columns
new_data <- new_data %>%
  dplyr::rename(
    Chela = chela,
    "PCR result" = "PCR Result",
    "Host_Tissue" = "Host Tissue"
  )

# See if we have any other mismatches
all_equal(new_data, full_data, ignore_col_order = FALSE, convert = TRUE)

# Convert START TIME column from character to POSIX
new_data$`START TIME` <- strptime(new_data$`START TIME`, format = '%m/%d/%Y %H:%M')

# See if we have any other mismatches
all_equal(new_data, full_data, ignore_col_order = FALSE, convert = TRUE)

# Join the 2016 data to the table with data from previous years
full_data <- bind_rows(full_data, new_data)
```

#### Read in 2017 data

```{r}
# Read in file from 2017
new_data <- read_excel(files[4], sheet = 2)

# Check if names match between 2017 data and our full_data table
all_equal(new_data, full_data, ignore_col_order = FALSE)

# Again, appears that the mismatches in column names are simple differences in name formatting, not overall sheet organization. Therefore we can simply rename our columns
new_data <- new_data %>%
  dplyr::rename(
    Chela = chela,
    "Host_Tissue" = "Host Tissue"
  )

# See if we have any other mismatches
all_equal(new_data, full_data, ignore_col_order = FALSE, convert = TRUE)

# Join the 2017 data to the table with data from previous years
full_data <- bind_rows(full_data, new_data)
```

#### Read in 2018 data

```{r}
# Read in file from 2018
new_data <- read_excel(files[5], sheet = 2)

# Check if names match between 2018 data and our full_data table
all_equal(full_data, new_data)

# Looks like there are 5 extra columns in the 2018 data. Let's see where these are
names(new_data)[names(full_data) != names(new_data)]

# Appears that in 2018 alternative definitions of maturity began to be examined. This resulted in extra columns for ln(CW), meas'd ln(Ch), calc ln(Ch), mat, and Mat1. Because there's some odd variable names there, remove based on indexing
new_data <- new_data[, -c(6:10)]

# Check again if names match between 2018 data and our full_data table
all_equal(full_data, new_data)

# Again, appears that the mismatches in column names are simple differences in name formatting, not overall sheet organization. Therefore we can simply rename our columns
new_data <- new_data %>%
  dplyr::rename(
    Chela = chela,
    "Host_Tissue" = "Host Tissue"
  )

# See if we have any other mismatches
all_equal(new_data, full_data, ignore_col_order = FALSE, convert = TRUE)

# Join the 2018 data to the table with data from previous years
full_data <- bind_rows(full_data, new_data)
```

#### Read in 2019 data

```{r}
# Read in file from 2019
new_data <- read_excel(files[6], sheet = 3)

# Check if names match between 2019 data and our full_data table
all_equal(full_data, new_data)

# Looks like there are 9 extra columns in the 2019 data. Let's see where these are
names(new_data)[names(full_data) != names(new_data)]

# Appears that the examination of alternative definitions of maturity continued. This resulted in six extra columns noting chela height and carapace width calculations. Furthermore, there are three columns on the end that should be removed - they appear to have been used only for calculations. Because there's some odd variable names there, remove based on indexing
new_data <- new_data[, -c(5, 7:11, 43:45)]

# Check again if names match between 2019 data and our full_data table
all_equal(full_data, new_data)

# Again, appears that the mismatches in column names are simple differences in name formatting, not overall sheet organization. Therefore we can simply rename our columns
new_data <- new_data %>%
  dplyr::rename(
    Chela = chela,
    "Host_Tissue" = "Host Tissue",
    VESSEL = Vessel,
    HAUL = Haul
  )

# See if we have any other mismatches
all_equal(new_data, full_data, ignore_col_order = FALSE, convert = TRUE)

# Join the 2019 data to the table with data from previous years
full_data <- bind_rows(full_data, new_data)
```

#### End of data import

#### Clean up column names so that none contain spaces

```{r}
names(full_data) <- str_replace_all(names(full_data), c(" " = "_", "," = ""))
```


We now have a single table (full_data) that contains approximately 7100 rows, with each row corresponding to the test results for a single crab. However, not all columns contain information relevant to the model. Therefore, we will now remove those columns from the data. We will remove the following columns for the following reasons:

SPNO: Gives the row number for that year, which we don't care about

Random: Contains NAs in all rows except 3

Collection_Comments: Checked, and no rows contain info relevant to model

ANOM_ID: Contains no useful information

General_Location: For all, is BS (Bering Sea)

Collected_By: Collector should be irrelevant

Reference_NO: Used for testing, irrelevant once PCR results obtained

DNA_Plate_No: Used for testing, irrelevant once PCR results obtained

DNA_Well_No: Used for testing, irrelevant once PCR results obtained

Host_Tissue: Either leg muscle or blood, should be irrelevant after PCR testing

Preservative: Irrelevant after PCR results obtained

Sample_Status: Irrelevant after PCR results obtained

C_V_H: Amalgamation of three other columns (CRUISE, VESSEL, and HAUL)

CRUISE: Survey leg is irrelevant once collection has been made

VESSEL: Also does not matter after collection

HAUL: Haul number of survey does not matter after collection

GEAR_DEPTH: Also have a column for BOTTOM DEPTH, and it would be quite surprising if any mid-water trawls picked up crab!

SURFACE_TEMP: Also have a column for bottom temp, which is more relevant for crab conditions

```{r}
full_data <- subset(full_data, select = -c(SPNO, Random, Collection_Comments, ANOM_ID, General_Location, Collected_By, Reference_NO, DNA_Plate_No, DNA_Well_No, Host_Tissue, Preservative, Sample_Status, C_V_H, CRUISE, VESSEL, HAUL, GEAR_DEPTH, SURFACE_TEMP))
```

#### Examining Additional Variables to Eliminate

```{r}
table(full_data$Egg_Cond)
# Can eliminate - indicates little beyond egg presence
full_data <- subset(full_data, select = -Egg_Cond)

table(full_data$Specific_Location)
# Can eliminate - sites are year-specific
full_data <- subset(full_data, select = -Specific_Location)

table(full_data$Egg_Color, full_data$PCR_result)
# Practically all eggs are either 0 (absent) or 4 (orange). We already have clutch fullness data which provides presence/absence information (and more), so this column can be removed.
full_data <- subset(full_data, select = -Egg_Color)

table(full_data$STRATUM, full_data$Year)
# Looks like for all of 2008, we have almost entirely NAs. Therefore, rather than losing a year, let's drop and use latitude/longitude. Since we're using latitude/longitude, we'll also drop STATIONID
full_data <- subset(full_data, select = -c(STRATUM, STATIONID))
```

#### Modifying Categorical Variables

We will now examine and eliminate unnecessary complexity by grouping variable categories and eliminating minimally-represented variation

```{r}
# First, we will convert START_TIME to Julian day (day from the start of the year)
full_data$JUL_DAY <- yday(full_data$START_TIME)
hist(full_data$JUL_DAY)
# We can now drop the START_TIME column
full_data <- subset(full_data, select = -START_TIME)

table(full_data$Sex)
sum(is.na(full_data$Sex))
# Eliminate all 4 rows where sex is unknown (=3), along with the 6 rows where sex is NA
full_data <- subset(full_data, Sex < 3)

table(full_data$Species_Name)
# Eliminate the one helmet crab included in the data set
full_data <- subset(full_data, Species_Name != "Telmessus cheiragonus")

table(full_data$Shell_Cond)
sum(is.na(full_data$Shell_Cond))
# #liminate all 0s, 1s, and 2s (soft-, fresh- and very very old-shell crab. 10 total among all 3 categories). Also eliminates the 10 crabs with NA shell condition
full_data <- subset(full_data, Shell_Cond < 5 & Shell_Cond > 1)

table(full_data$Clutch)
# Given the relatively small number of crabs with eggs, I think it makes more sense to change this to 0 (no eggs present) and 1 (eggs present). Essentially, all crabs with egg codes from 1-6 will be converted to egg code 1
full_data$Clutch <- recode_factor(full_data$Clutch, "0" = "0", "1" = "1", "2" = "1", "3" = "1", "4" = "1", "5" = "1", "6" = "1")
# Check if we have any males with a listed clutch fullness
table(full_data$Clutch, full_data$Sex)
# Assuming the 3 with no chela heights are females accidentally entered as males. The one with a chela height listed will be removed regardless later, as it has no PCR result associated with it, so we'll just change all to females
full_data[!is.na(full_data$Clutch) & full_data$Sex == 1, ]$Sex <- 2
table(full_data$Clutch, full_data$Sex)

table(full_data$PCR_result)
# Only want to keep conclusive PCR results, so eliminating all NAs and 3s
full_data <- subset(full_data, PCR_result < 2)

table(full_data$Year)
```

#### Eliminating NAs

We eliminated some NAs above, but want to examine in more detail to see how many are feasible to eliminate

```{r}
colSums(is.na(full_data))

# Looks like the only NAs are for chela height and clutch condition. NAs in these categories are expected - chela height is not measured for females and clutch size is not measured for females. However, let's verify in more detail

colSums(is.na(full_data[full_data$Sex == 1, ]))
colSums(is.na(full_data[full_data$Sex == 2, ]))

# Looks like we do have NAs for chela height for males and NAs for clutch condition for females. We'll sort those out later, when we create sex-specific dataframes

# Otherwise, looks like we've already sorted out our NAs!
```

#### Checking for Data Entry Errors for Continuous Variables

```{r}
# Since male and female Chionoecetes crab have quite different sizes, we'll check for each of these separately
hist(full_data[full_data$Sex == 1, ]$Size)
hist(full_data[full_data$Sex == 2, ]$Size)
summary(full_data[full_data$Sex == 1, ]$Size)
summary(full_data[full_data$Sex == 2, ]$Size)
# Looks like no data entry errors for size!

# Now moving on to chela height
hist(full_data$Chela)
summary(full_data$Chela)
# Check to see if any females have non-NA entries for chela height
table(full_data[full_data$Sex == 2, ]$Chela)
# Looks like we have 7 crab! Assume these are males who were accidentally mislabeled as females - coming up with a specific chela height seems unlikely as a possible data entry error

full_data[!is.na(full_data$Chela) & full_data$Sex == 2, ]

# Data entry errors for chela height have now been fixed

# Now moving on to latitude and longitude. For this, we will map each point
world <- ne_countries(scale = "medium", returnclass = "sf")
ggplot(data = world) +
  geom_sf() +
  geom_point(data = full_data, aes(x = START_LONGITUDE, y = START_LATITUDE), size = 1, shape = 23, fill = "darkred")

# No global data entry errors - all are clustered within the EBS region. However, let's examine on a smaller scale
ggplot(data = world) +
  geom_sf() +
  geom_point(data = full_data, aes(x = START_LONGITUDE, y = START_LATITUDE), size = 1, shape = 23, fill = "darkred") +
  coord_sf(xlim = c(-180, -160), ylim = c(50, 70), expand = FALSE)

# From this map, it appears that our points fall into 6 distinct clusters - Bristol Bay, Pribilof Islands, outer shelf south, St. Matthew Island, outer shelf north, and St. Lawrence Island. This indicates we may want to modify this into a categorical variable rather than a continuous one
# We will address that in the chunk below

# Moving on to bottom depth
hist(full_data$BOTTOM_DEPTH)
summary(full_data$BOTTOM_DEPTH)
# No data entry errors apparent

# Moving on to bottom temperature
hist(full_data$Bottom_Temp)
summary(full_data$Bottom_Temp)
# A temperature of 9 degrees seems anomalous - let's examine further
head(sort(unique(full_data$Bottom_Temp), decreasing = TRUE))
# It certainly is high, but not high enough to conclude a data entry error was made

# Moving on to Julian day
hist(full_data$JUL_DAY)
# No data entry errors apparent - all seem within standard survey date ranges
```

#### Adding Categorical Variable for Area from Latitude-Longitude

As described above, our points fall into 6 distinct clusters. They are as follows (north to south): 

- Bristol Bay (BB)
- Pribilof Islands (PI)
- South Slope (SS)
- St. Matthew Island (SMI)
- North Slope (NS)
- St. Lawrence Island (SLI)

To have the option of defining location as one categorical variable rather than two continuous variables (latitude and longitude), we will create a new categorical variable. 

They will be assigned numbers from 1-6, following the same order as above.

```{r}
# Plot overall map of haul locations
ggplot(data = world) +
  geom_sf() +
  geom_point(data = full_data, aes(x = START_LONGITUDE, y = START_LATITUDE), size = 1, shape = 23, fill = "darkred") +
  coord_sf(xlim = c(-180, -160), ylim = c(50, 70), expand = FALSE)

## Plot same map but with lines showing boundaries of Bristol Bay cluster (BB)
ggplot(data = world) +
  geom_sf() +
  geom_point(data = full_data, aes(x = START_LONGITUDE, y = START_LATITUDE), size = 1, shape = 23, fill = "darkred") +
  geom_hline(yintercept = 58, linetype = "solid", color = "black") +
  geom_vline(xintercept = -167, linetype = "solid", color = "black") +
  coord_sf(xlim = c(-180, -160), ylim = c(50, 70), expand = FALSE)

# The Bristol Bay cluster will be assigned as trawls with latitude < 58 and longitude > -167

# Create new column for location (initially 0 for all crab)
full_data$Location <- "Not Assigned"

# Assign "BB" to the Bristol Bay cluster
full_data[full_data$START_LATITUDE < 57 & full_data$START_LONGITUDE > -167,]$Location <- "BB"

# Moving on to Pribilof Islands

ggplot(data = world) +
  geom_sf() +
  geom_point(data = full_data, aes(x = START_LONGITUDE, y = START_LATITUDE), size = 1, shape = 23, fill = "darkred") +
  geom_hline(yintercept = 59, linetype = "solid", color = "black") +
  geom_vline(xintercept = -167, linetype = "solid", color = "black") +
  geom_vline(xintercept = -173, linetype = "solid", color = "black") +
  coord_sf(xlim = c(-180, -160), ylim = c(50, 70), expand = FALSE)

# The Pribilof Islands cluster will be assigned as trawls with latitude < 59 and longitude between -167 and -173

full_data[full_data$START_LATITUDE < 59 & full_data$START_LONGITUDE > -173 & full_data$START_LONGITUDE < -167,]$Location <- "PI"

# Moving on to South Slope

ggplot(data = world) +
  geom_sf() +
  geom_point(data = full_data, aes(x = START_LONGITUDE, y = START_LATITUDE), size = 1, shape = 23, fill = "darkred") +
  geom_hline(yintercept = 59.5, linetype = "solid", color = "black") +
  geom_vline(xintercept = -173, linetype = "solid", color = "black") +
  coord_sf(xlim = c(-180, -160), ylim = c(50, 70), expand = FALSE)

# The South Slope cluster will be assigned as trawls with latitude < 59.5 and longitude < -173

full_data[full_data$START_LATITUDE < 59.5 & full_data$START_LONGITUDE < -173,]$Location <- "SS"

# Moving on to St. Matthew Island

ggplot(data = world) +
  geom_sf() +
  geom_point(data = full_data, aes(x = START_LONGITUDE, y = START_LATITUDE), size = 1, shape = 23, fill = "darkred") +
  geom_hline(yintercept = 59, linetype = "solid", color = "black") +
  geom_hline(yintercept = 62, linetype = "solid", color = "black") +
  geom_vline(xintercept = -167, linetype = "solid", color = "black") +
  geom_vline(xintercept = -173, linetype = "solid", color = "black") +
  coord_sf(xlim = c(-180, -160), ylim = c(50, 70), expand = FALSE)

# The St. Matthew Island cluster will be assigned as trawls with latitude between 59 and 62 and longitude between -167 and -173

full_data[full_data$START_LATITUDE > 59 & full_data$START_LATITUDE < 62 & full_data$START_LONGITUDE > -173 & full_data$START_LONGITUDE < -167,]$Location <- "SMI"

# Moving on to North Slope

ggplot(data = world) +
  geom_sf() +
  geom_point(data = full_data, aes(x = START_LONGITUDE, y = START_LATITUDE), size = 1, shape = 23, fill = "darkred") +
  geom_hline(yintercept = 61, linetype = "solid", color = "black") +
  geom_hline(yintercept = 63, linetype = "solid", color = "black") +
  geom_vline(xintercept = -173, linetype = "solid", color = "black") +
  coord_sf(xlim = c(-180, -160), ylim = c(50, 70), expand = FALSE)

# The North Slope cluster will be assigned as trawls with latitude between 61 and 63 and longitude < -173

full_data[full_data$START_LATITUDE > 61 & full_data$START_LATITUDE < 63 & full_data$START_LONGITUDE < -173,]$Location <- "NS"

# Moving on to St. Lawrence Island

ggplot(data = world) +
  geom_sf() +
  geom_point(data = full_data, aes(x = START_LONGITUDE, y = START_LATITUDE), size = 1, shape = 23, fill = "darkred") +
  geom_hline(yintercept = 63, linetype = "solid", color = "black") +
  coord_sf(xlim = c(-180, -160), ylim = c(50, 70), expand = FALSE)

# The St. Lawrence Island cluster will be assigned as trawls with latitude > 63

full_data[full_data$START_LATITUDE > 63,]$Location <- "SLI"

# Alright, let's check that we did this correctly

# See if any hauls were not assigned a location code, along with the distribution of crabs
table(full_data$Location)

# Looks like all crabs were assigned a location! Let's check to ensure that each was placed in the correct location by mapping

ggplot(data = world) +
  geom_sf() +
  geom_point(data = full_data, aes(x = START_LONGITUDE, y = START_LATITUDE, color = Location), size = 1, shape = 16) +
  coord_sf(xlim = c(-180, -160), ylim = c(50, 70), expand = FALSE)

# Fantastic, looks great!
```

#### Modifying Variable Names

We now have fully cleaned our full data. For the sake of consistency, however, we will modify our variable names, which have inconsistent notation and capitalization

```{r}
# See current variable names
names(full_data)

# Modify names to be consistent
names(full_data) <- c("Species", "Sex", "Carapace_Width", "Chela_Height", "Shell_Condition", "Clutch_Fullness", "PCR_Result", "Year", "Latitude", "Longitude", "Depth", "Bottom_Temp", "Julian_Day", "Location")

names(full_data)
```

#### Creating sex-specific table for males

Since the effect of chela height and clutch fullness may be interesting, we will create two sub-tables. One will only include males with a measurement for chela height, and the other will only include females with a measurement for clutch fullness.

```{r}
# Create table containing data for males
males <- full_data[full_data$Sex == 1,]

# Drop clutch fullness from males dataset
males <- subset(males, select = -Clutch_Fullness)

# See how many males have an NA for chela height and if any other NAs are present
colSums(is.na(males))

# Drop the 851 males without a value for chela height from the dataset
males <- na.omit(males)
```

#### Creating sex-specific table for females

This table will, as described above, only include females with a measurement for clutch fullness

```{r}
# Create table containing data for females
females <- full_data[full_data$Sex == 2, ]

# Drop chela height from females dataset
females <- subset(females, select = -Chela_Height)

# See how many females have an NA for clutch fullness and if any other NAs are present
colSums(is.na(females))

# Drop the 341 females without a value for clutch fullness from the dataset
females <- na.omit(females)
```

#### Eliminate sex-specific variables from our dataset with both male and female crabs

Since any analysis that includes all crabs will include neither chela height or clutch fullness (since those measurements are, definitionally, only available for either male or female crab), we will eliminate both of these variables from our dataset that includes all crabs

```{r}
# Remove sex-specific variables
all_crabs <- subset(full_data, select = -c(Chela_Height, Clutch_Fullness))

# Check if we have any NAs
colSums(is.na(all_crabs))
```

#### Congratulations, we finished with data cleaning! Now it's time to write our datasets to a table

At this point, we have three data frames - one including only males, one including only females, and one with both males and females.

Along with these, we will write two more to files - one for each of our two species, _Chionoecetes bairdi_ and _Chionoecetes opilio_

```{r}
# Write males to file
write.csv(males, file = "../output/NOAA_EBS_trawl_survey/modified_data/cleaned_males.csv", row.names = FALSE)

# Write females to file
write.csv(females, file = "../output/NOAA_EBS_trawl_survey/modified_data/cleaned_females.csv", row.names = FALSE)

# Write all crabs to file
write.csv(all_crabs, file = "../output/NOAA_EBS_trawl_survey/modified_data/cleaned_all_crabs.csv", row.names = FALSE)

# Write C. bairdi to file
write.csv(all_crabs[all_crabs$Species == "Chionoecetes bairdi", ],
          file = "../output/NOAA_EBS_trawl_survey/modified_data/cleaned_bairdi.csv",
          row.names = FALSE)

# Write C. opilio to file
write.csv(all_crabs[all_crabs$Species == "Chionoecetes opilio", ],
          file = "../output/NOAA_EBS_trawl_survey/modified_data/cleaned_opilio.csv",
          row.names = FALSE)
```


Addendum: Creating a quick graph of PCR positive results among all crabs

```{r}
crab.sum <- all_crabs %>%
  group_by(Year, Species) %>%
  summarise(
    pct_inf = mean(PCR_Result)
  )

ggplot(data = crab.sum,
       aes(x = Year, y = pct_inf, fill = Species)) +
  geom_bar(stat = "identity", position = position_dodge()) +
  scale_fill_brewer(palette = "Set1") +
  ylab("% Infected By Hematodinium (PCR)")
   
```
