==> scripts/2_3_fixing_temp_file_structure.ipynb <==
In the previous script, we cleaned our Tanner crab data. Now, it's time to address temperature. The temperature files are stored in some wildly disparate file structures, alternating often between years and occasionally within a single year. 

Our goal is as follows:

- Ensure all files are the raw .txt files directly from the data logger. This will require eliminating numerous .dtf files, along with some Excel summary files

- Change all file structures to the following format: year/survey/leg/.txt files

- Change all .txt file names to be Tidbit number.txt (note:tidbit is the name of the data logger type). When possible, those Tidbit numbers will equal those in data/ADFG_SE_AK_pot_surveys/Pot_Set_Data_for_Tanner_and_RKC_surveys.csv

- Standardize all folder names. Ex: Tidbit 1 from Leg 2 of the 2015 RKC survey would be in 2015/RKC/Leg_2/1.txt

- Standardize names of skipper files (which match pots and Tidbit data). Standard will be simply be skipper_data, and it'll be inside the relevant folder

When an already-amalgamated file contains both the skipper and Tidbit info, we'll simply use that! We'll call it skipper_tidbit_data.xls We'll fix the file structure too though, just to be consistent. If we just have Tidbit data, we'll use that too - we'll call it tidbit_data.xls

Summary of file structures is as follows. Note: this is approximate, as some years have different structures for the two surveys (one may be Tanner -> LegNum.txt, the other may be RKC -> Leg -> num.txt)

Year -> Crab -> Leg -> num.txt
2011
2012
2013
2014
2015
2016
2017
2018
2019

Year -> CrabLeg -> num.txt
2006
2007
2008
2009
2010
2017


Year -> num,date.txt
2005

# Moving Files to Output

To ensure we've got a raw, original copy of all our data within the data/ directory, we'll copy the whole folder over to the output/ directory. Note that only *some* files within our directory of temperature data will be modified (in name, in contents, or both) over the next few scripts. But it's a whoooole lot easier to just read in all files from a single directory instead of constantly switching back and forth, so that's what we'll do!

pwd

# Move all files over to new directory
!cp -r ../data/ADFG_SE_AK_pot_surveys/Tidbits/ ../output/ADFG_SE_AK_pot_surveys/cleaned_data/

# Rename new directory for clarity
!mv ../output/ADFG_SE_AK_pot_surveys/cleaned_data/Tidbits/ ../output/ADFG_SE_AK_pot_surveys/cleaned_data/temperature_data/

# 2005

Note: temperature data is only available for the Tanner crab survey. 

Dates for each leg are as follows:
Leg 1: Oct. 7th-14th
Leg 2: Oct. 20th-26th

Dates on the filenames appear to be dates of download for the Tidbit files. Therefore, all with Oct19 in the filename are from Leg 1, and all with Oct29 in the filename are from Leg 2.

cd ../output/ADFG_SE_AK_pot_surveys/cleaned_data/temperature_data/2005

!ls
# We want the skipper data and the .txt files
# Everything else can be tossed

!rm *.dtf
!rm *.jmp
!rm *.mdb
!rm *.doc
!rm 05_Tanner_Tidbit_Data.xls

mkdir Tanner_survey

# Move all .txt files to the Tanner_survey folder
!mv *.TXT Tanner_survey

# Rename skipper data form, move into Tanner survey folder
!mv 05_Tanner_Survey_Skipper_Data.xls Tanner_survey/skipper_data.xls
!mv 05_Tanner_Survey_Tidbit_and_Skipper_Data_Access_Prep.xls Tanner_survey/skipper_tidbit_data.xls

# Move into that Tanner_survey folder
%cd Tanner_survey/

!ls

# Create new directories for each leg
!mkdir Leg_1
!mkdir Leg_2

# Move all files with Oct19 into Leg_1
!mv *Oct19*.TXT Leg_1

# Move into the Leg 1 folder
%cd Leg_1

# Rename all files to strip everything except the number and the .txt
# Ex: T5,Oct192005.TXT -> 5.txt
# We just have four files, so instead of setting up a Bash script, we'll do them individually
!mv T5,Oct192005.TXT 5.txt
!mv T7,Oct192005.TXT 7.txt
!mv T10,Oct192005.TXT 10.txt
!mv T11,Oct192005.TXT 11.txt

ls

# Move out to encompassing folder
%cd ..

ls

# Move all .txt files to Leg 2 folder
!mv *.TXT Leg_2/

%cd Leg_2

# Rename all files to strip everything except the number and the .txt
# We just have five files, so instead of setting up a Bash script, we'll do them individually
!mv T5,Oct292005.TXT 5.txt
!mv T7,Oct292005.TXT 7.txt
!mv T9,Oct292005.TXT 9.txt
!mv T10,Oct292005.TXT 10.txt
!mv T11,Oct292005.TXT 11.txt

# Done! That was a tricky one. Moving out to the 2006 folder now
%cd ../../../2006

# 2006

Alright, this one will be simpler

!ls

# Remove the following files and directories
!rm -r Shrimp_Survey_Leg_1_2006/
!rm *.tmp
!rm *.JMP
!rm *.jmp
!rm 06_Shrimp*.xls
!rm *.doc
!rm Temperature_database.xls
!rm Tidbit_template.xls
!rm T11.txt
!rm 06_RKC_Leg_1,2,3_Tidbit_Data.xls
!rm 06_Tanner_Survey_L_1,2_Tidbit_Data.xls

# Make new directories for each survey type
!mkdir Tanner_survey
!mkdir RKC_survey

# Move folders and files into correct directories, renaming along way
!mv RKC_*Leg* RKC_survey
!mv RKC_survey_leg_1_2006/ RKC_survey
!mv Tanner_survey_leg* Tanner_survey
!mv 06_Tanner_Survey_Skipper_Data.xls Tanner_survey/skipper_data.xls
!mv 06_RKC_Leg_1,2,3_Skipper_Data.xls RKC_survey/skipper_data.xls
!mv 06_Tanner_Survey_L_1,2_Access_Prep.xls Tanner_survey/skipper_tidbit_data.xls
!mv 06_RKC_Leg_1,2,3_Skipper_and_Tidbit_Joined_Access_prep.xls RKC_survey/skipper_tidbit_data.xls

%cd RKC_survey/

# Rename files for each leg
!mv RKC_survey_leg_1_2006/ Leg_1
!mv RKC_Survey_Leg_2_2006/ Leg_2
!mv RKC_Survey_Leg_3_2006/ Leg_3

# Remove .dtf files in each leg folder, along with an Excel doc
!rm */*.dtf
!rm */*.DTF
!rm */*.xls

# Leg 1: Need to remove the T prefix, then all good
%cd Leg_1/
!rename 's/T//' T*.TXT
%cd ..


# Leg 2 and 3 still have issues
# Ex: Leg 2, Tidbit 1 is T1T22006.
# Need to remove the first T, and everything after the 2nd T
# Note: will use rename command. Not part of base Linux, but install with sudo apt install rename

# Fix Leg 2
%cd Leg_2
!rename 's/T206//' *T206.TXT
!rename 's/T22006//' *T22006.TXT
!rename 's/T//' T*.TXT
%cd ../Leg_3


# Fix Leg 3
!rename 's/T306//' *T306.TXT
!rename 's/T//' T*.TXT

# Moving on to Tanner data for 2006
%cd ../../Tanner_survey/

# Rename folders
!mv Tanner_survey_leg_1_2006/ Leg_1/
!mv Tanner_survey_leg_2_2006/ Leg_2/

# Remove .dtf files
!rm */*.dtf

Done with 2006! Moving on to 2007

# 2007


%cd ../../2007

!ls

# Remove the following files and directories
!rm -r Shrimp_*
!rm 07_Shrimp_*
!rm 06_Shrimp_*
!rm *.JMP
!rm *.jmp
!rm 07_RKC_Survey_L_1,_2,_3_Tidbit_Data.xls

# Make new directories for each survey type
!mkdir Tanner_survey
!mkdir RKC_survey

# Move folders and files into correct directories
!mv RKC_Survey_Leg* RKC_survey
!mv Tanner_*_2007/ Tanner_survey
!mv 07_Tanner_L_1,2_Skipper_Data.xls Tanner_survey/skipper_data.xls
!mv 07_RKC_L_1,_2,_3_Skipper_Data.xls RKC_survey/skipper_data.xls
!mv 07_RKC_Survey_L_1,_2,_3_Tidbit_and_Skipper_Joined_Access_Prep.xls RKC_survey/skipper_tidbit_data.xls
!mv 07_Tanner_L_1,2_Tidbit_Data.xls Tanner_survey/tidbit_data.xls

%cd RKC_survey/

# Rename files for each leg
!mv RKC_Survey_Leg_1_2007/ Leg_1
!mv RKC_Survey_Leg_2_2007/ Leg_2
!mv RKC_Survey_Leg_3_2007/ Leg_3

# Remove .dtf files in each leg folder, along with an Excel doc
!rm */*.dtf
!rm */*.xls

# Legs 1, 2, and 3 all have various issues. Renaming to standardize
# Note: will use rename command. Not part of base Linux, but install with sudo apt install rename

# The 20_07 tidbit isn't found in the database. However, it is present in the comments of the 2007 pot set data
# A number of pot sets with Tidbit 20 have in the comments "Tidbit 20-07", "TB 2007", or some variant
# We'll put that as a note for things to change when merging the full data
%cd Leg_1/
!mv 20_07.TXT 20-07.txt
!rename 's/_07//' *_*.TXT


# Fix Leg 2
%cd ../Leg_2/
!rename 's/_62807//' *_62807.TXT
!rename 's/62907//' *62907.TXT
!rename 's/62807//' *62807.TXT
!rename 's/T//' T*.TXT
# Again, we'll just rename 20_07 to 20-07
!mv 20_07.TXT 20-07.txt


# Fix Leg 3
%cd ../Leg_3/
!rename 's/71907//' *71907.TXT
!rename 's/T//' T*.TXT

# Moving on to Tanner data
%cd ../../Tanner_survey/

# Rename folders
!mv Tanner_survey_leg_1_2007/ Leg_1/
!mv Tanner_Survey_Leg_2_2007/ Leg_2/

# Remove .dtf files
!rm */*.dtf

# Change all _ to - to better fit the format of the file structure overall
%cd Leg_1
!rename 's/_/-/' *.TXT

%cd ../Leg_2
!rename 's/_/-/' *.TXT

# Alright, done! On to 2008
%cd ../../../2008/

# 2008

!ls

# Remove the following files and directories
!rm -r Shrimp_*
!rm -r GKC_Observer_Program_08/
!rm *.JMP
!rm *.jmp
!rm 08_Shrimp_*
!rm *emplate_for_Access_*.xls
!rm 2008_RKC_Skipper_Data.xls
!rm RKC_2008_Linked_temp_skipper.xls
# Have a Tidbit file in this directory, but it's from September, when neither the RKC or Tanner surveys were happening. Remove it
!rm 22_07.txt

# Make new directories for each survey type
!mkdir Tanner_survey
!mkdir RKC_survey

# Move folders and files into correct directories
!mv RKC_Survey_Leg* RKC_survey
!mv Tanner_Survey_Leg*/ Tanner_survey
!mv 08_Tanner_Survey_Leg_1,_2_Skipper_Data.xls Tanner_survey/skipper_data.xls
!mv 2008_RKC_Skipper_Data_for_Access_4.xls RKC_survey/skipper_data.xls
!mv 08_RKC_Survey_Leg_1,_2,_3,_Access_Prep.xls RKC_survey/skipper_tidbit_data.xls
!mv 08_Tanner_Survey_Leg_1,_2_Access_Prep.xls Tanner_survey/skipper_tidbit_data.xls
!mv 08_Tanner_Survey_Leg_1,_2_Tidbit_Data.xls Tanner_survey/tidbit_data.xls
!mv 2008_RKC_Tidbit_Data_for_Access.xls RKC_survey/tidbit_data.xls

%cd RKC_survey/

# Rename files for each leg
!mv RKC_Survey_Leg_1_2008/ Leg_1
!mv RKC_Survey_Leg_2_2008/ Leg_2
!mv RKC_Survey_Leg_3_2008/ Leg_3

# Remove .dtf files in each leg folder, along with an Excel doc
!rm */*.dtf
!rm */*.xls

# Leg 1 and 2 have a summary .txt file that we'll remove
!rm Leg_1/RKC_L1_2008.TXT
!rm Leg_2/RKC_2008_L2.TXT

# We'll also remove the underscores from 20_07 to make it 2007
# Finally, we'll remove the ending _07 from all other files ending in *_07.txt
# This is because in the pot set data file, the same logger is called, say, 18, 187, and 1807 at different points,
# despite evidently being one data logger with one file. To keep it simple, we'll just refer to it as Tidbit 18.
!rename 's/_07.TXT/07.txt/' Leg_*/20_07.TXT
!rename 's/_07.TXT/.txt/' Leg_*/*_07.TXT

# Moving on to Tanner data
%cd ../Tanner_survey/

# Rename folders
!mv Tanner_survey_Leg_1_2008/ Leg_1/
!mv Tanner_Survey_Leg_2_2008/ Leg_2/

# Remove .dtf files
!rm */*.dtf

# Leg 1 needs the _07 removed from all except 20_07.txt
%cd Leg_1

# Change 20_07 to 2007 to match the formatting in the pot set data (data/ADFG_SE_AK_pot_surveys/Pot_Set_Data_for_Tanner_and_RKC_surveys.csv)
!mv 20_07.TXT 2007.txt
!rename 's/_07//' *_07.TXT

# Alright, now for Leg 2 of the Tanner crab survey
%cd ../Leg_2

!rename 's/1008//' *1008.TXT
!rename 's/T//' T*.TXT
!rename 's/07.TXT/.TXT/' *07.TXT
!mv 710.TXT 7.TXT
!mv 200710.TXT 2007.TXT
!mv 240710.TXT 24.TXT

# Alright, done! On to 2009
%cd ../../../2009/

# 2009

!ls

# Remove the following files and directories
!rm -r Shrimp_*
!rm -r GKC_Observer_Program_09/
!rm -r Pilot_Project_Lynn_Sisters__09_Temp_Data/
!rm *.jmp
!rm 09_Shrimp_Survey_Leg_1,2_Temp_Data.xls
# Looked up RKC_1B and it took place between legs of the RKC survey, with 
# no matching entries in the skipper/tidbit data. Therefore removing
!rm -r RKC_1B

# Make new directories for each survey type
!mkdir Tanner_survey
!mkdir RKC_survey

# Move folders and files into correct directories
!mv RKC_L* RKC_survey
!mv Tanner_L*/ Tanner_survey
!mv 09_Tanner_Survey_Leg_1,2_Skipper_Data.xls Tanner_survey/skipper_data.xls
!mv 09_RKC_Survey_Leg_1,2,3_Skipper_Data.xls RKC_survey/skipper_data.xls
!mv 09_Tanner_Survey_Leg_1,2_Skipper_and_Temp_Joined_Access_Prep.xls Tanner_survey/skipper_tidbit_data.xls
!mv 09_RKC_Survey_Leg_1,2,3_Skipper_and_Temp_Joined_Access_Prep.xls RKC_survey/skipper_tidbit_data.xls
!mv 09_Tanner_Survey_Leg_1,2_Temp_Data.xls Tanner_survey/tidbit_data.xls
!mv 09_RKC_Survey_Leg_1,2,3_Temp_Data.xls RKC_survey/tidbit_data.xls

%cd RKC_survey/

# Rename files for each leg
!mv RKC_L1_09/ Leg_1
!mv RKC_L2_09/ Leg_2
!mv RKC_L3_09/ Leg_3

# Remove .dtf files in each leg folder
!rm */*.dtf

# Leg 2 has a summary .xls file that we'll remove
# Otherwise, no other weird files!
!rm Leg_2/RKC_L2_09.xls

# Legs 2 and 3 are fine, but Leg 1 has some weird prefixes
# Each file is prefixed by "L1_"
%cd Leg_1
!rename 's/L1_//' L1_*
# Worked for almost all, but one has a _ instead of a _
!mv _5.TXT 5.TXT

# Moving on to Tanner data
%cd ../../Tanner_survey/

# Rename folders
!mv Tanner_L1_09/ Leg_1/
!mv Tanner_L2_09/ Leg_2/

# Remove .dtf files and .hobo files
!rm */*.dtf
!rm */*.hobo

# Alright, done! On to 2010
%cd ../../2010/

# 2010

!ls

# Remove the following files and directories
!rm -r Shrimp_*
!rm -r GKC_Observer_Program*
!rm -r NPRB_RKC_Temps/ # from November, much after the RKC survey. Presumed NPRB special project
!rm *.jmp

# Make new directories for each survey type
!mkdir RKC_survey
!mv Tanner_L1,2_10/ Tanner_survey/

# Move folders and files into correct directories
!mv RKC_L*_10/ RKC_survey
!mv Tanner_L1,2_Skipper_Data.xls Tanner_survey/skipper_data.xls
!mv RKC_L2,3_Skipper_Data.xls RKC_survey/skipper_data.xls
!mv Tanner_L1,2_Tidbit_Data.xls Tanner_survey/tidbit_data.xls
!mv RKC_L2,3_Tidbit_Data.xls RKC_survey/tidbit_data.xls

%cd RKC_survey/

# Rename files for each leg
!mv RKC_L1_10/ Leg_1
!mv RKC_L2_10/ Leg_2
!mv RKC_L3_10/ Leg_3

# Remove .dtf, .jmp, and .hobo files in each leg folder
!rm */*.dtf
!rm */*.hobo
!rm */*.jmp

# We've got some weird files here
# Leg 1 has a separate skipper and temp data sheet that we'll remove
# It also has a merged skipper_tidbit_data sheet that we'll move out to the main folder
!mv Leg_1/10_RKC_Survey_Leg_1_Temp_and_Skipper_Joined.xlsx skipper_tidbit_data_leg_1.xls
!rm */*.xls

# Moving on to Tanner data
%cd ../Tanner_survey/

# Remove .dtf files and .hobo files
!rm *.dtf
!rm *.hobo

# Files contain data from both legs of survey, so we're done!
# On to 2011
%cd ../../2011/

# 2011

!ls

# Remove the following directories
!rm -r Shrimp/
!rm -r NPRB_RKC/
!rm -r GKC/

# We've already got separate directories for each survey type
# Rename for consistency
!mv RKC/ RKC_survey/
!mv Tanner/ Tanner_survey/

%cd RKC_survey/

# Rename the legs of the survey for consistency
!mv Leg_I/ Leg_1
!mv Leg_II/ Leg_2
!mv Leg_III/ Leg_3

# Remove .hobo, .jmp, and .dtf files from each directory
!rm */*.hobo
!rm */*.dtf
!rm */*.jmp

# skipper and tidbit data files were created
# However, they weren't merged between legs. Currently inside
# each leg's directory. Move out to main directory now

# Leg 1
!mv Leg_1/11_RKC_Survey_Leg_1_Skipper_Data.xls skipper_data_leg_1.xls
!mv Leg_1/11_RKC_Survey_Leg_1_Temp_Data.xls tidbit_data_leg_1.xls

# Leg 2
!mv Leg_2/11_RKC_Survey_Leg_2_Skipper_Data.xls skipper_data_leg_2.xls
!mv Leg_2/11_RKC_Survey_Leg_2_Temp_Data.xls tidbit_data_leg_2.xls

# Leg 3
!mv Leg_3/11_RKC_Survey_Leg_3_Skipper_Data.xls skipper_data_leg_3.xls
!mv Leg_3/11_RKC_Survey_Leg_3_Temp_Data.xls tidbit_data_leg_3.xls

# Moving on to Tanner data
%cd ../Tanner_survey/

# Remove .dtf, .jmp, and .hobo files
!rm */*.dtf
!rm */*.hobo
!rm */*.jmp

# skipper and tidbit data files were created
# However, they weren't merged between legs. Currently inside
# each leg's directory. Move out to main directory now

# Leg 1
!mv Leg_1/2011_Tanner_Leg_I_Skipper.xls skipper_data_leg_1.xls
!mv Leg_1/2011_Tanner_Leg_I_Temp.xls tidbit_data_leg_1.xls

# Leg 2
!mv Leg_2/2011_Tanner_Leg_2_Skipper_Data.xls skipper_data_leg_2.xls
!mv Leg_2/2011_Tanner_Leg_2_Temp_Data.xls tidbit_data_leg_2.xls

# Alright, done! On to 2012
%cd ../../2012/

# 2012

!ls

# Remove the following directories
!rm -r Shrimp/
!rm -r NPRB/
!rm -r GKC/
!rm -r Yakutat_Dungeness/

# We've already got separate directories for each survey type
# Rename for consistency
!mv RKC/ RKC_survey/
!mv Tanner/ Tanner_survey/

%cd RKC_survey/

# Rename the legs of the survey for consistency
!mv Leg_I/ Leg_1
!mv Leg_II/ Leg_2
!mv Leg_III/ Leg_3

# Remove .hobo, .jmp, and .dtf files from each directory
!rm */*.hobo
!rm */*.dtf
!rm */*.jmp

# skipper and tidbit data files were created
# However, they weren't merged between legs. Currently inside
# each leg's directory. Move out to main directory now

# Leg 1
!mv Leg_1/12_RKC_Skipper_Data_Leg_I.xls skipper_data_leg_1.xls
!mv Leg_1/12_RKC_Temp_Data_Leg_I.xls tidbit_data_leg_1.xls

# Leg 2
!mv Leg_2/12_RKC_Skipper_Data_Leg_II.xlsx skipper_data_leg_2.xls
!mv Leg_2/12_RKC_Temp_Data_Leg_II.xlsx tidbit_data_leg_2.xls

# Leg 3
!mv Leg_3/12_RKC_Skipper_Data_Leg_III.xlsx skipper_data_leg_3.xls
!mv Leg_3/12_RKC_Temp_Data_Leg_III.xlsx tidbit_data_leg_3.xls

# Moving on to Tanner data
%cd ../Tanner_survey/

# Rename folders for consistency
!mv Leg_I/ Leg_1/
!mv Leg_II/ Leg_2/

# Remove .dtf, and .hobo files
!rm */*.dtf
!rm */*.hobo

# skipper and tidbit data files were created
# However, they weren't merged between legs. Currently inside
# each leg's directory. Move out to main directory now
# Also, filtered temp files were created, evidently with a bar for 
# water temp. Remove, since we can do that ourselves

# Leg 1
!mv Leg_1/2012_Tanner_Leg_I_Skipper_Data.xlsx skipper_data_leg_1.xls
!mv Leg_1/2012_Tanner_Leg_I_Temp_Data.xlsx tidbit_data_leg_1.xls
!rm Leg_1/2012_Tanner_Leg_I_Temp_filtered.xlsx

# Leg 2
!mv Leg_2/2012_Tanner_Leg_II_skipper_data.xlsx skipper_data_leg_2.xls
!mv Leg_2/2012_Tanner_Leg_II_Temp.xls tidbit_data_leg_2.xls
!rm Leg_2/2012_Tanner_Leg_II_Temp_filtered.xlsx

# Alright, done! On to 2013
%cd ../../2013/

# 2013

!ls

# Remove the GKC directory
!rm -r GKC/

# We've already got separate directories for each survey type
# Rename for consistency
!mv RKC/ RKC_survey/
!mv Tanner/ Tanner_survey/

%cd RKC_survey/

# Rename the legs of the survey for consistency
!mv Leg_I/ Leg_1
!mv Leg_II/ Leg_2
!mv Leg_III/ Leg_3

# Remove .hobo and .dtf files from each directory
!rm */*.hobo
!rm */*.dtf

# skipper and tidbit data files were created
# Rename them for consistency 

!mv 2013_RKC_pot_data.xlsx skipper_data.xls
!mv 2013_RKC_temp.xlsx tidbit_data.xls

# Also an unnecessary tidbit file in Leg 1, remove
!rm Leg_1/2013_RKC_temp_Leg_I.xlsx

# Leg 1 is fine
# Legs 2 and 3 have zeroes prefixed to all single-digit numbers
# Remove these

%cd Leg_2/
!rename 's/0//' 0?.txt
!rename 's/0//' 0?.TXT

# Moving to Leg 3
%cd ../Leg_3/
!rename 's/0//' 0?.txt
!rename 's/0//' 0?.TXT

# Moving on to Tanner data
%cd ../../Tanner_survey/

# Looks like loggers were just downloaded once, at the end of Leg 2
# Remove .hobo files
!rm *.hobo

# Rename skipper and temp data files for consistency
!mv 2013_Tanner_pot_data.xlsx skipper_data.xls
!mv 2013_Tanner_temp_data.xlsx tidbit_data.xls

# Alright, done! On to 2014
%cd ../../2014/

# 2014

!ls

# Remove the GKC directory
!rm -r GKC/

# We've already got separate directories for each survey type
# Rename for consistency
!mv RKC/ RKC_survey/
!mv Tanner/ Tanner_survey/

%cd RKC_survey/

# Rename the legs of the survey for consistency
!mv Leg_I/ Leg_1
!mv Leg_II/ Leg_2
!mv Leg_III/ Leg_3

# Remove .hobo files from each directory
!rm */*.hobo

# Tidbit data files were created
# Rename them for consistency 

!mv Leg_1/2014_RKC_Leg_I_temp_summary.xlsx tidbit_data_leg_1.xls
!mv Leg_2/2014_leg_II_RKC_temp_summary.xlsx tidbit_data_leg_2.xls
!mv Leg_3/2014_Leg_III_RKC_temp_summary.xlsx tidbit_data_leg_3.xls

# Legs have zeroes prefixed to some single-digit numbers
# Remove these

%cd Leg_1/
!rename 's/0//' 0?.txt

# Move to Leg 2
%cd ../Leg_2/
!rename 's/0//' 0?.txt


# Moving to Leg 3
%cd ../Leg_3/
!rename 's/0//' 0?.txt

# Moving on to Tanner data
%cd ../../Tanner_survey/

# Rename folders to match consensus
!mv Leg_I/ Leg_1/

# Remove .hobo files
!rm */*.hobo

# Rename temp data files for consistency
!mv Leg_1/2014_TC_leg_I_temp_summary.xlsx tidbit_data_leg_1.xls
!mv Leg_2/2014_TC_leg_II_temp_summary.xlsx tidbit_data_leg_2.xls

# Legs have zeroes prefixed to some single-digit numbers
# Remove these

%cd Leg_1/
!rename 's/0//' 0?.txt

# Move to Leg 2
%cd ../Leg_2/
!rename 's/0//' 0?.txt

# Alright, done! On to 2015
%cd ../../../2015/

# 2015

!ls

# We've already got separate directories for each survey type
# Rename for consistency
!mv RKC_Temp_2015/ RKC_survey/
!mv 2015_Tanner_Tidbit_Data/ Tanner_survey/

%cd RKC_survey/

# Rename the legs of the survey for consistency
!mv Leg_I/ Leg_1
!mv Leg_II/ Leg_2
!mv Leg_III/ Leg_3

# Remove .hobo files from each directory
!rm */*.hobo

# Tidbit data files were created
# Rename them for consistency 

!mv Leg_1/2015_JNU_and_Barlow_temperature.xlsx tidbit_data_leg_1.xlsx
!mv Leg_2/2015_Deadmans_Reach,_Excursion_Inlet,_and_Lynn_Sisters_temperature.xlsx tidbit_data_leg_2.xlsx
!mv Leg_3/2015_Pybus_Bay,_Gambier_Bay,_and_Seymour_Canal_temperature.xls tidbit_data_leg_3.xls

# Legs have zeroes prefixed to some single-digit numbers
# Remove these

%cd Leg_1/
!rename 's/0//' 0?.txt

# Move to Leg 2
%cd ../Leg_2/
!rename 's/0//' 0?.txt

# Moving to Leg 3
%cd ../Leg_3/
!rename 's/0//' 0?.txt

# Moving on to Tanner data
%cd ../../Tanner_survey/

# Rename folders to match consensus
!mv Trip_1/ Leg_1/
!mv Trip_2 Leg_2/

# Remove .hobo files
!rm */*.hobo

# Folder contains some filtered data files
# Remove for consistency, since we'll just do that ourselves
!rm Leg_1/2015_Tanner_temperature_leg_I_filtered.xlsx
!rm Leg_2/2015_Tanner_leg_2_temperature_data_filtered.xlsx

# Rename temp data files for consistency
!mv Leg_1/2015_Tanner_temperature_leg_I.xlsx tidbit_data_leg_1.xlsx
!mv Leg_2/2015_Tanner_leg_2_temperature_data.xlsx tidbit_data_leg_2.xlsx

# Legs have zeroes prefixed to some single-digit numbers
# Remove these, along with some suffixes in Leg 2

%cd Leg_1/
!rename 's/0//' 0?.txt

# Move to Leg 2
%cd ../Leg_2/
!rename 's/0//' 0?.txt
!rename 's/_0//' *_0.txt
!rename 's/0//' 0?.txt


# Alright, done! On to 2016
%cd ../../../2016/

# 2016

!ls

# Remove irrelevant folders and files
!rm -r RKC_MR_temp_excursion/
!rm Thumbs.db
!rm Juneau_2016_Temp_Graph.JPG
!rm RKC_AllTempGraphs_2016.docx
!rm RKC_LegIII_Temperature_Graphs.docx
!rm RKCS_SkipperData_2016.xlsx

# We've already got separate directories for each survey type
# Rename for consistency
!mv RKC_temp_data/ RKC_survey/
!mv 2016_Tanner_Tidbit_Data/ Tanner_survey/

%cd RKC_survey/

# Rename the legs of the survey for consistency
!mv RKC_Leg1_2016/ Leg_1
!mv RKC_Leg2_2016/ Leg_2
!mv RKC_Leg3_2016/ Leg_3

# Remove unnecessary file
!rm RKC2.7z

# Remove .hobo files from each directory
!rm */*.hobo

# Remove other unnecessary files
!rm Leg_1/Juneau_2016_Temp_Graph.JPG
!rm Leg_1/Thumbs.db
!rm Leg_1/Leg1_Temperature_Data_2016.xls
!rm Leg_3/*_Details.txt

# Tidbit data files were created
# Rename them for consistency 

!mv Leg_1/Leg1_Temperature_Data.xls tidbit_data_leg_1.xls
!mv Leg_1/Leg1_SkipperData.xlsx skipper_data_leg_1.xlsx

!mv Leg_2/RKC_LegII_Temp_Data_2016.xlsx tidbit_data_leg_2.xlsx

!mv Leg_3/RKC_Leg3_AllTidbitData.xlsx tidbit_data_leg_3.xlsx

# Legs have zeroes prefixed to some single-digit numbers
# Remove these

%cd Leg_1/
!rename 's/0//' 0?.xls

# Move to Leg 2
%cd ../Leg_2/
!rename 's/0//' 0?.csv

# Moving to Leg 3
%cd ../Leg_3/
!rename 's/0//' 0?.csv

# Moving on to Tanner data
%cd ../../Tanner_survey/

# Rename folders to match consensus
!mv Trip_1/ Leg_1/
!mv TCS_Leg_II_Tidbit_Data/ Leg_2/

# Remove .hobo files
!rm */*.hobo
!rm */*_Details.txt

# Legs have zeroes prefixed to some single-digit numbers
# Remove these, along with some suffixes in Leg 2

%cd Leg_1/
!rename 's/0//' 0?.txt

# Move to Leg 2
%cd ../Leg_2/
!rename 's/_0//' *_0.txt
!rename 's/0//' 0?.txt


# Alright, done! On to 2017
%cd ../../../2017/

# 2017

!ls

# We've already got separate directories for each survey type
# Rename for consistency
!mkdir RKC_survey
!mv 2017_Tanner_leg_1/ Tanner_survey/
!mv 2017_*/ RKC_survey/

%cd RKC_survey/

# Rename the legs of the survey for consistency
!mv 2017_rkc_leg_1/ Leg_1
!mv 2017_rkc_leg_2/ Leg_2
!mv 2017_RKC_Leg3/ Leg_3


# Remove .hobo files from each directory
!rm */*.hobo

# Remove other unnecessary files
!rm Leg_*/*_Details.txt
!rm Leg_3/19_0_Details_0.txt

# Legs have zeroes prefixed to some single-digit numbers
# Remove these

%cd Leg_1/
!rename 's/0//' 0?.txt

# Move to Leg 2
%cd ../Leg_2/
!rename 's/_0//' *_0.txt
!rename 's/0//' 0?.txt

# Moving to Leg 3
%cd ../Leg_3/
!rename 's/_0//' *_0.txt
!rename 's/0//' 0?.txt
!mv 09_0.txt 9_0.txt

# Moving on to Tanner data
%cd ../../Tanner_survey/

# Rename folders to match consensus
!mv Leg1/ Leg_1/

# Remove .hobo files
!rm */*.hobo

# Legs have zeroes prefixed to some single-digit numbers
# Remove these, along with some suffixes in Leg 2

%cd Leg_1/
!rename 's/_0//' *_0.txt
!rename 's/0//' 0?.txt


# Alright, done! On to 2018
%cd ../../../2018/

# 2018

!ls

# We've already got separate directories for each survey type
# Rename for consistency
!mv 2018_Tanner_Tidbit_Data/ Tanner_survey/
!mv RKC_2018/ RKC_survey/

%cd RKC_survey/

# Rename the legs of the survey for consistency
!mv Leg_I/ Leg_1
!mv legII/ Leg_2
!mv Leg_III/ Leg_3


# Remove .hobo files from each directory
!rm */*.hobo

# Legs have zeroes prefixed to some single-digit numbers
# Remove these

%cd Leg_1/
!rename 's/_0//' *_0.txt
!rename 's/0//' 0?.txt

# Move to Leg 2
%cd ../Leg_2/
!rename 's/_0//' *_0.txt
!rename 's/0//' 0?.txt

# Moving to Leg 3
%cd ../Leg_3/
!rename 's/_0_0//' *_0.txt
!rename 's/0//' 0?.txt

# Moving on to Tanner data
%cd ../../Tanner_survey/

# Rename folders to match consensus
!mv 2018_tanner_leg_2/ Leg_2

# Remove .hobo files
!rm */*.hobo

# Legs have zeroes prefixed to some single-digit numbers
# Remove these, along with some suffixes in Leg 2

%cd Leg_1/
!rename 's/0//' 0?.txt

%cd ../Leg_2/
!rename 's/_0//' *_0.txt
!rename 's/0//' 0?.txt


# Alright, done! On to 2019
%cd ../../../2019/

# 2019

# We've already got separate directories for each survey type
# Rename for consistency
!mv RKC/ RKC_survey/

%cd RKC_survey/

# Rename the legs of the survey for consistency
!mv 2019_RKC_Leg_I/ Leg_1
!mv 2019_RKC_LEG_2/ Leg_2
!mv 2019_RKC_Leg3/ Leg_3


# Extracting .hobo files from Leg 2
Sadly, here's where we need to step away from reproducibility for a bit (dammit proprietary software!). Data files for Leg 2 of the 2019 RKC survey are only available in .hobo format. Therefore, we need to install the HOBOware software (available at https://www.onsetcomp.com/hoboware-free-download/) and manually convert them from .hobo to .csv files

I already did the conversion, putting the files in data/ADFG_SE_AK_pot_surveys/Tidbits/2019/RKC/2019_RKC_LEG_2. If you need to do them yourself, go to Tools -> Bulk File Export -> Select Files To Export, and select all .hobo files from data/ADFG_SE_AK_pot_surveys/Tidbits/2019/RKC/2019_RKC_LEG_2/. This will convert all files to .csv files. You can then go back to this script and continue.

# Remove .hobo files from each directory
!rm */*_Details.txt
!rm */*.hobo

# Remove the export description file created due to our bulk transport
!rm Leg_2/Export_*.log

# Legs have zeroes prefixed to some single-digit numbers, plus some suffixes to adjust

%cd Leg_1/
!rename 's/0//' 0?.txt
!mv 20_0.txt 20.txt

%cd ../Leg_2/
!rename 's/_0//' *_0.csv
!rename 's/0//' 0?.csv

%cd ../Leg_3/
!rename 's/_0//' *_0.txt
!rename 's/0//' 0?.txt


pwd

# Alright, done! Back to main folder!
%cd ../../../

# Main Folder

# Remove unnecessary files and folders
!rm -r Access_template/
!rm -r BOXCAR3/
!rm -r HOBOware/
!rm -r Temp_folder_for_Chris/
!rm GKC_Avg._Temps.JNB
!rm Tidbit_template.xls
!rm *.JNB
!rm -r External_data/
!rm Linking_Temperature_and_skipper_data.doc

# Done!

We now have a pretty dang solid set of temperature files, along with some descriptor files, all within a consistent file structure. Go team!
