==> scripts/2_7_cleaning_full_data.Rmd <==
---
title: "2_7_cleaning_full_data.Rmd"
author: "Aspen Coyle"
date: "8/4/2022"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

## Introduction

In the previous script, we finished by creating a full dataframe containing temperature and catch data for each crab (excluding crabs without either data on their temperature or the presence/absence of Bitter Crab Syndrome).

Now it's time to start exploring the data! Let's check out what's been going on with these crabs, and see the relationship to temperature.

#### Load libraries (and install if necessary), and load packages

```{r libraries, message=FALSE, warning=FALSE}
# Add all required libraries here
list.of.packages <- c("tidyverse", "lubridate", "beepr")
# Get names of all required packages that aren't installed
new.packages <- list.of.packages[!(list.of.packages %in% installed.packages()[, "Package"])]
# Install all new packages
if(length(new.packages)) install.packages(new.packages)


# Load all required libraries
lapply(list.of.packages, FUN = function(X) {
  do.call("require", list(X))
})
```

# Read in data and fix it up

```{r}
# Examine things related to crabs (1 row per crab sampled)
crab_dat <- read.csv(file = "../output/ADFG_SE_AK_pot_surveys/cleaned_data/crab_data/BCS_examined_crab_with_temperature.csv")

# Examine things related to location (1 row per pot deployment)
pot_dat <- read.csv(file = "../output/ADFG_SE_AK_pot_surveys/cleaned_data/pot_data_with_temperature.csv")
```


# Exploring the Pot Data

```{r}
#### Temperature and Site --------------------------
exper <- pot_dat %>%
  group_by(Location) %>%
  summarize(temp_avg  = mean(temp),
            temp_sd = sd(temp),
            n_pots = n()) %>%
  ungroup()

ggplot(exper) +
  geom_bar(aes(x = Location, y = temp_avg),
           stat = "identity") +
  geom_errorbar(aes(x = Location, ymin = temp_avg - temp_sd, ymax = temp_avg + temp_sd)) +
  theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust = 1))

# Alright, neat little graph here! Looks like most sites have a broadly similar range of temperatures, with some exceptions. Port Camden in particular seems to be substantially higher than practically every other site!

# Let's look at how temperature relates to year!
exper <- pot_dat %>%
  group_by(Year) %>%
  summarize(temp_avg  = mean(temp),
            temp_sd = sd(temp),
            n_pots = n()) %>%
  ungroup()

ggplot(exper) +
  geom_bar(aes(x = Year, y = temp_avg),
           stat = "identity") +
  geom_errorbar(aes(x = Year, ymin = temp_avg - temp_sd, ymax = temp_avg + temp_sd)) +
  theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust = 1))

#### Temp vs. time of year ----------------------------
# We'll use Julian day for this
pot_dat$Jul.Day <- as.Date(pot_dat$Time.Hauled)
pot_dat$Jul.Day <- yday(pot_dat$Jul.Day)

# First, a scatterplot of all temps and Julian days
ggplot(pot_dat, aes(x = Jul.Day, y = temp)) +
  geom_point()

# Alright, we can see an overall gentle curve, along with the timing of each survey (Tanner vs. RKC)

# Now, let's make a line plot!
exper <- pot_dat %>%
  group_by(Jul.Day) %>%
  summarize(temp_avg = mean(temp))

ggplot(exper, aes( x= Jul.Day, y = temp_avg)) +
  geom_line() +
  geom_point()

#### Temp vs. depth ---------------------------
# We'll look at the average depth of each site vs. the average temp
exper <- pot_dat %>%
  group_by(Location) %>%
  summarize(depth_avg  = mean(Depth.Fathoms),
            temp_avg = mean(temp),
            temp_sd = sd(temp),
            depth_sd = sd(Depth.Fathoms),
            n_pots = n()) %>%
  ungroup()

# First, depth of each site
ggplot(exper) +
  geom_bar(aes(x = Location, y = depth_avg),
           stat = "identity") +
  geom_errorbar(aes(x = Location, ymin = depth_avg - depth_sd, ymax = depth_avg + depth_sd)) +
  theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust = 1))

# Now, depth vs. temperature
ggplot(exper, aes(x = depth_avg, y = temp_avg)) +
  geom_point()

# Now let's switch it up to maximum depth (more of an examination of the water bodies) and temperature
exper <- pot_dat %>%
  group_by(Location) %>%
  summarize(depth_max  = max(Depth.Fathoms),
            temp_avg = mean(temp),
            temp_sd = sd(temp),
            n_pots = n(),
            avg_day = mean(Jul.Day),
            day_sd = sd(Jul.Day))%>%
  ungroup()

# First, depth of each site
ggplot(exper) +
  geom_bar(aes(x = Location, y = depth_max),
           stat = "identity") +
  theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust = 1))

# Now, depth vs. temperature
ggplot(exper, aes(x = depth_max, y = temp_avg)) +
  geom_point()

# Hmm, doesn't seem to be a super strong relationship there. Neat! Note though, this is averaged by site

# What about average date each site was sampled?
ggplot(exper) +
  geom_bar(aes(x = Location, y = avg_day),
           stat = "identity") +
  geom_errorbar(aes(x = Location, ymin = avg_day - day_sd, ymax = avg_day + day_sd)) +
  theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust = 1))  

# Hmm, alright. Looks like a few of the sites are specific to the later survey, while most aren't. 
# Let's verify this with a table
table(pot_dat$Location, pot_dat$Project)

# Interesting! Alright, looks like surveys fall into the following buckets:
# Tanner crab only:
#       - Glacier Bay
#       - Icy Strait
#       - Port Camden
#       - Stephens Passage
#       - Thomas Bay
# Both Tanner and RKC survey:
#       - Hoklham Bay
# RKC only:
#       - All other surveys

# That might complicate things a bit - disentangling the effect of location and date - but hey, ah well.

# One thing I'm concerned about is a strong relationship between temperature and date. Let's figure that out
cor(pot_dat
    $temp, pot_dat$Jul.Day)
# Alright, that's acceptable! Let's just graph it too
ggplot(pot_dat, aes(x = Jul.Day, y = temp)) +
  geom_point()
# Clearly a warming trend over time, but it does flatten by October (realistically, peaks and goes back down)

# Let's also double check on latitude. Almost definitely fine, because our spatial scale is small, but hey
ggplot(pot_dat, aes(x = Latitude.Decimal.Degrees, y = temp)) +
  geom_point()

# Ooh, that's not good, we've got a few pots with crazy outlier values. Alright, let's switch over to cleaning up our crab data, since that's what we really care about!
```

#### Cleaning crab data

The plan is to go through each column one by one and clean up the dataframe in that order

```{r}
# Look at the columns to be examined
names(crab_dat)

# Remove irrelevant columns
crab_dat <- crab_dat %>%
  dplyr::select(-c(Trip.Start.Date, Trip.End.Date))

# Check each column for NAs
colSums(is.na(crab_dat))

# Alright, some NAs in columns are understandable (male crab won't have egg condition values, for instance), others aren't. 

# Let's remove NA values in columns where we care about them

# Sex
crab_dat <- crab_dat[!is.na(crab_dat$Sex), ]

# Width
crab_dat <- crab_dat[!is.na(crab_dat$Width.Millimeters), ]

# Shell condition
crab_dat <- crab_dat[!is.na(crab_dat$Shell.Condition), ]

# Leg condition
crab_dat <- crab_dat[!is.na(crab_dat$Leg.Condition), ]

# Depth
crab_dat <- crab_dat[!is.na(crab_dat$Depth.Fathoms), ]

# Check again
colSums(is.na(crab_dat))

# Alright, only remaining NAs are in male-specific columns (chela height) or female-specific (egg-related)

#### Renaming Columns ---------------------------------------
# A lot of the names of our variables aren't super descriptive or follow a good pattern. Let's fix that!

crab_dat <- crab_dat %>%
  dplyr::rename(c(
    Survey = Project,
    Leg = Trip.No,
    Site = Location,
    CW = Width.Millimeters,
    Chela.Ht = Chela.Height.Millimeters,
    Latitude = Latitude.Decimal.Degrees,
    Longitude = Longitude.Decimal.Degrees,
    Jul.Day = Time.Hauled,         # Converting this to Julian day
    Depth = Depth.Fathoms,
    Temp = temp
  ))

# Quickly shift Time.Hauled column to Julian day
crab_dat$Jul.Day <- as.Date(crab_dat$Jul.Day)
crab_dat$Jul.Day <- yday(crab_dat$Jul.Day)



# Proceed with examining each column specifically in detail

#### Year ---------------------------------------
table(crab_dat$Year)
min(crab_dat$Year)
max(crab_dat$Year)
# Looks all good, unsurprisingly!

#### Survey --------------------------------------
table(crab_dat$Survey)
# Also unsurprisingly, all good!

#### Leg -----------------------------------
table(crab_dat$Leg)
# Also looks good!

#### Site --------------------------
table(crab_dat$Site)
# We've got some very low numbers here at some sites
# Cape Fanshaw has three and Round Rock has one. We'll remove both
crab_dat <- crab_dat %>%
  dplyr::filter(Site != "Cape_Fanshaw" & Site != "Round_Rock")
# Alright, now all good!

#### Pot.No ----------------------------
# At this point, we don't actually care about pot number - that was just a way to join it with the pot data. Let's remove it!
crab_dat <- crab_dat %>%
  dplyr::select(-Pot.No)

#### Sex --------------------------------
table(crab_dat$Sex)
# We'll switch the data codes over to the ones used by ADFG.
# 1 = Male, 2 = Female
crab_dat[crab_dat$Sex == "Male", ]$Sex <- 1
crab_dat[crab_dat$Sex == "Female", ]$Sex <- 2

crab_dat$Sex <- as.factor(crab_dat$Sex)

#### Carapace Width (CW) -------------------------------
head(sort(crab_dat$CW), n = 10)
# Upon sorting and looking at the full table, all these look legit except the first one. No way there's a mature female with eggs and a CW of 9. Let's remove that
crab_dat <- crab_dat %>%
  filter(CW > 9)

head(sort(crab_dat$CW, decreasing = TRUE), n = 10)
# Also looks all good. Let's check our females specifically, since they mature at smaller sizes
head(sort(crab_dat[crab_dat$Sex == 2, ]$CW, decreasing = TRUE), n = 10)

#### Chela Height ------------------------------
head(sort(crab_dat$Chela.Ht), n = 10)
# I'm almost certain using chela height is going to be a total waste of time, and I'm not planning on including it in the final model. Just keeping it in to be safe. Still, it's implausible to me that you can get a measurement of a crab with a chela height <= 3mm
crab_dat <- crab_dat %>%
  mutate(Chela.Ht = replace(Chela.Ht, Chela.Ht <= 3, NA))

# Also change all chela heights from females with NA, since it's only biologically meaningful on males (chela height:CW ratio marks maturity)
crab_dat <- crab_dat %>%
  mutate(Chela.Ht = replace(Chela.Ht, Sex == 2, NA))
# All good!

#### Recruit.Status ----------------------------
table(crab_dat$Recruit.Status)
# Looks fine to me!

#### Shell Condition
table(crab_dat$Shell.Condition)
# Hmm, we've only got 12 shell condition 1s. That's not a whole lot, let's remove them
crab_dat <- crab_dat %>%
  filter(Shell.Condition != 1)
# All good!

#### Egg.Condition ---------------------------------------
table(crab_dat$Egg.Condition)
# To be safe, turn any males into NAs
crab_dat <- crab_dat %>%
  mutate(Egg.Condition = replace(Egg.Condition, Sex == 1, NA))
# Looks fine!

#### Egg.Development ------------------------------------------------------
table(crab_dat$Egg.Development)
# Replace any values for males with NAs
crab_dat <- crab_dat %>%
  mutate(Egg.Development = replace(Egg.Development, Sex == 1, NA))

#### Leg.Condition -------------------------------------------------
table(crab_dat$Leg.Condition, useNA = "always")
# Change to either 0 (not damaged) or 1 (damaged)
crab_dat[crab_dat$Leg.Condition == 1, ]$Leg.Condition <- 0
crab_dat[crab_dat$Leg.Condition != 0, ]$Leg.Condition <- 1
table(crab_dat$Leg.Condition, useNA = "always")
# Looks fine!

#### Bitter -------------------------------------
table(crab_dat$Bitter)
# Looks fine!

#### Egg.Percent -------------------------------
table(crab_dat$Egg.Percent)
# We have some odd values in here, but hey, who cares if they didn't round. We'll just treat this as numeric later on

#### Blackmat -----------------------------------
table(crab_dat$Blackmat)

#### Latitude ----------------------------------------
#Let's eliminate everything with a latitude < 50 or > 60, anything else is outside the boundaries of the survey realistically
crab_dat <- crab_dat %>%
  filter(Latitude > 50) %>%
  filter(Latitude < 60)
# Looks good

#### Longitude ------------------------------------------------
# Let's do the same with longitude, eliminating all with a latitude < -140 or > -130
crab_dat <- crab_dat %>%
  filter(Longitude > -140) %>%
  filter(Longitude < -130)
# Looks good

#### Day ---------------------------------------------------
min(crab_dat$Jul.Day)
# That's June 4th (or 3rd, in leap years). Reasonable!
max(crab_dat$Jul.Day)
# That's Oct. 30th or 31st. Also reasonable!

#### Depth --------------------------------------------------------
head(sort(crab_dat$Depth))
head(sort(crab_dat$Depth, decreasing = TRUE))
# Both values look realistic. Depth is in fathoms, so 48 feet to ~1150 feet isn't too crazy. Still, let's look at a few years just to be sure
crab_dat %>%
  group_by(Year) %>%
  summarise(max = max(Depth))
# Yep, broadly similar maximum depths across surveys!

#### Temp --------------------------------------------------
head(sort(crab_dat$Temp))
head(sort(crab_dat$Temp, decreasing = TRUE))
# Looks good to me! Unsurprising though, we already fixed up temp earlier when reviewing our pot data and looking for outlier temperatures.
```

# Modifying Temperature to Average By Year and Location

```{r}
# Group temperature data by year and location
crab_dat <- crab_dat %>%
  group_by(Site, Year) %>%
  mutate(temp_avg = mean(Temp)) %>%
  ungroup()

# Drop the original temp column
crab_dat <- crab_dat %>%
  select(-Temp)

# For consistency, rename avg_temp to Temp
crab_dat <- crab_dat %>%
  dplyr::rename(
    Temp = temp_avg)
```


Alright, all crab data is now cleaned! Amazing. Let's write out that file. We didn't make any major edits, so let's just overwrite the one we made previously
```{r}

write.csv(crab_dat, file = "../output/ADFG_SE_AK_pot_surveys/cleaned_data/crab_data/BCS_examined_crab_with_temperature.csv", row.names = FALSE)
beep(8)
```

# Understanding Temperature-Bitterness Relationship

Let's also make a quick preliminary graph looking at the effects of temperature and its impact on bitterness

```{r}
bitter <- crab_dat
bitter$Temp <- round(bitter$Temp, digits = 1)
table(bitter$Temp)


bitter <- bitter %>%
  group_by(Temp) %>%
  summarize(bitter_avg = mean(Bitter)) %>%
  ungroup()

ggplot(bitter) +
  geom_point(aes(x = Temp, y = bitter_avg))

# Okay, looks like there's an overall negative relationship between temperature and bitterness. That's really weird because in my final model, there's a positive relationship. I also created a model with just temperature + random effects, and there was still a positive relationship. So let's check those random effects to see if there's some confounding issue there!

# Site 

bitter <- crab_dat
bitter$Temp <- round(bitter$Temp, digits = 0)
table(bitter$Temp)

bitter <- bitter %>%
  group_by(Temp, Site) %>%
  summarize(bitter_avg = mean(Bitter)) %>%
  ungroup()

ggplot(bitter) +
  geom_jitter(aes(x = Temp, y = bitter_avg, col = Site)) +
  geom_smooth(aes(x = Temp, y = bitter_avg, col = Site), method = lm, se = FALSE)

# Year

bitter <- crab_dat
bitter$Temp <- round(bitter$Temp, digits = 0)
bitter$Year <- as.factor(bitter$Year)

bitter <- bitter %>%
  group_by(Temp, Year) %>%
  summarize(bitter_avg = mean(Bitter)) %>%
  ungroup()

ggplot(bitter) +
  geom_jitter(aes(x = Temp, y = bitter_avg, col = Year)) +
  geom_smooth(aes(x = Temp, y = bitter_avg, col = Year), method = lm, se = FALSE)

# Alright, looks like site is the issue! Let's look at the rates of bitterness by site

bitter <- crab_dat

bitter <- bitter %>%
  group_by(Site) %>%
  summarize(bitter_avg = mean(Bitter)) %>%
  ungroup()

ggplot(bitter) +
  geom_bar(aes(x = Site, y = bitter_avg), stat = "identity") +
  theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1))

# And just to check, rates of bitterness by year

bitter <- crab_dat

bitter <- bitter %>%
  group_by(Year) %>%
  summarize(bitter_avg = mean(Bitter)) %>%
  ungroup()

ggplot(bitter) +
  geom_bar(aes(x = Year, y = bitter_avg), stat = "identity") +
  theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1))
```


Alright, so we seem to have a situation where within each site, warmer waters = more infected crab. However, warmer sites tend to have lower rates of infection in the first place. Weird! Let's confirm that by making some site-specific graphs of infection rates

#### Temperature Per Site

```{r fig.height=50}

# Copy over data, rounding temp to nearest 0.1 degree
bitter <- crab_dat
bitter$Temp <- round(bitter$Temp, digits = 0.1)

# Create filepath to save graphs
temp_graphs <- "../output/ADFG_SE_AK_pot_surveys/diagnostic_images/site_specific_graphs/infection_rates/without_year"

# Create a vector of all sites
sites <- unique(bitter$Site)

# Start our for loop. Each time we will filter data to only include data from one site, summarize, and graph
for (i in 1:length(sites)) {
  site_dat <- bitter %>%
    dplyr::filter(Site == sites[i])
  
  site_avgs <- site_dat %>%
    group_by(Temp) %>%
    summarize(bitter_avg = mean(Bitter),
              n_crabs = n())
  
  ggplot(site_avgs) +
    geom_point(aes(x = Temp, y = bitter_avg, size = n_crabs)) +
    geom_smooth(aes(x = Temp, y = bitter_avg, weight = n_crabs), method = lm, se = FALSE) +
    labs(title = sites[i]) +
    xlim(3, 9.5) +
    ylim(0, 1)
  
  ggsave(filename = paste0(temp_graphs, sites[i], ".png"), plot = last_plot(), device = png())
  dev.off()
}

bitter <- crab_dat
bitter$Temp <- round(bitter$Temp, digits = 1)

site_avgs <- bitter %>%
  group_by(Temp, Site) %>%
  summarize(bitter_avg = mean(Bitter),
            n_crabs = n())

ggplot(site_avgs) +
  geom_point(aes(x = Temp, y = bitter_avg, size = n_crabs)) +
  geom_smooth(aes(x = Temp, y = bitter_avg, weight = n_crabs), method = lm, se = FALSE) +
  facet_wrap(vars(Site), ncol = 3)

ggsave(filename = "../output/ADFG_SE_AK_pot_surveys/diagnostic_images/site_specific_graphs/all_sites.png")


```

#### Temperature Per Site Per Year
```{r fig.height=30}


#### Now we'll do the same thing again, but adding color for year!

# Copy over data, rounding temp to nearest 1 degree
bitter <- crab_dat
bitter$Temp <- round(bitter$Temp, digits = 1)
# Set year to factor
bitter$Year <- as.factor(bitter$Year)

# Create filepath to save graphs
temp_graphs <- "../output/ADFG_SE_AK_pot_surveys/diagnostic_images/site_specific_graphs/infection_rates/with_year/"

# Create a vector of all sites
sites <- unique(bitter$Site)

# Start our for loop. Each time we will filter data to only include data from one site, summarize, and graph
for (i in 1:length(sites)) {
  site_dat <- bitter %>%
    dplyr::filter(Site == sites[i])
  
  site_avgs <- site_dat %>%
    group_by(Temp, Year) %>%
    summarize(bitter_avg = mean(Bitter),
              n_crabs = n())
  
  ggplot(site_avgs) +
    geom_point(aes(x = Temp, y = bitter_avg, col = Year, size = n_crabs)) +
    geom_smooth(aes(x = Temp, y = bitter_avg, col = Year, weight = n_crabs), method = lm, se = FALSE) +
    labs(title = sites[i]) +
    xlim(3, 9.5) +
    ylim(0, 1)
  
  ggsave(filename = paste0(temp_graphs, sites[i], ".png"), plot = last_plot(), device = png())
  dev.off()
}
```

```{r fig.height= 30}

### Plot infection data as binomial with regression line
# This is the same basic process as before, but just with one change
# Rather than plotting the data as averaged, we'll plot it binomially with a regression line

# Copy over data. No need to round temp this time.
bitter <- crab_dat

# Create filepath to save graphs
temp_graphs <- "../output/ADFG_SE_AK_pot_surveys/diagnostic_images/site_specific_graphs/logistic_regression/"

# Create a vector of all sites
sites <- unique(bitter$Site)

# Start our for loop. Each time we will filter data to only include data from one site and graph
for (i in 1:length(sites)) {
  site_dat <- bitter %>%
    dplyr::filter(Site == sites[i])
  
  ggplot(site_dat, aes(x = Temp, y = Bitter)) +
    geom_point(alpha = 0.5) +
    stat_smooth(method = "lm", se = FALSE, method.args = list(family = binomial))
  
  ggsave(filename = paste0(temp_graphs, sites[i], ".png"), plot = last_plot(), device = png())
  dev.off()
}


ggplot(crab_dat, aes(x = Temp, y = Bitter)) +
  geom_jitter() +
  stat_smooth(method = "glm", se = FALSE, method.args = list(family = binomial)) +
  facet_wrap(vars(Site), ncol = 3)

ggsave(filename = "../output/ADFG_SE_AK_pot_surveys/diagnostic_images/site_specific_graphs/logistic_regression/all_sites.png")
```


[Trash]

Miscellaneous quick functions meant to examine survey design
```{r}
table(pot_dat$Year, pot_dat$Project)

table(pot_dat$Trip.Start.Date, pot_dat$Project)
table(pot_dat$Trip.End.Date, pot_dat$Project)

rkc_dat <- pot_dat %>%
  filter(Project == "Red King Crab Survey")

table(rkc_dat$Year, rkc_dat$Location)

tan_dat <- pot_dat %>%
  filter(Project == "Tanner Crab Survey")

table(tan_dat$Year, tan_dat$Location)

table(pot_dat$Location, pot_dat$Project)

```
