==> scripts/3_1_data_exploration.Rmd <==
---
title: "3_1_data_exploration"
author: "Aspen Coyle"
date: "11/8/2023"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```


#### Load libraries (and install if necessary), and load packages

```{r libraries, message=FALSE, warning=FALSE}
# Add all required libraries here
list.of.packages <- c("tidyverse", "lubridate", "beepr", "maps", "sf", "ggmap", "leaflet", "ggthemes", "cowplot")
# Get names of all required packages that aren't installed
new.packages <- list.of.packages[!(list.of.packages %in% installed.packages()[, "Package"])]
# Install all new packages
if(length(new.packages)) install.packages(new.packages)


# Load all required libraries
lapply(list.of.packages, FUN = function(X) {
  do.call("require", list(X))
})


```

## Introduction

In this script, we'll be doing some data examination to get a bit more detail on the results of our model and the overall patterns in our data


## Graphing Crab Density vs. Infection Rates

Here, we'll create a graph of the density of crabs per site per year vs. the rate of infection


```{r density}
# Read in data
crab_dat <- read.csv(file = "../output/ADFG_SE_AK_pot_surveys/cleaned_data/crab_data/BCS_cleaned.csv", header = TRUE)

# Drop all except year, project, location, pot number, infection status
crab_dat <- crab_dat %>%
  select("Year", "Project", "Location", "Pot.No", "Bitter")

# Group by year and location, get number of crabs and number of pots per site
crab_dat <- crab_dat %>%
  group_by(Year, Location)%>%
  summarise(infection_rate = (sum(Bitter)/length(Bitter)),
            num_crabs = length(Bitter),
            num_pots = max(Pot.No))

# Create new column for crab density
crab_dat <- crab_dat %>%
  mutate(crab_density = num_crabs/num_pots)

# Plot crab density vs. infection rate
ggplot(data = crab_dat, aes(x = crab_density, y = infection_rate)) +
  geom_point()+
  geom_smooth(method = "lm")

# Save plot
ggsave(filename = "../output/ADFG_SE_AK_pot_surveys/diagnostic_images/density_vs_infection.png")
```

### Graph of infection rates by year

```{r}
# Read in data
crab_dat <- read.csv(file = "../output/ADFG_SE_AK_pot_surveys/cleaned_data/crab_data/BCS_cleaned.csv", header = TRUE)

# Drop all except year, infection status
crab_dat <- crab_dat %>%
  dplyr::select("Year", "Bitter")

# Group by year and location, get frequency of bitter crab in each year
crab_dat <- crab_dat %>%
  group_by(Year)%>%
  summarise(infection_rate = (sum(Bitter)/length(Bitter)))

# Create line graph using ggplot to visualize infection rate changes over time

crab_dat %>%
  ggplot(aes(x = Year, y = infection_rate)) +
  labs(y = italic("Rate of Infection"),
       x = "Year") +
  geom_line()
         
         
         +
  geom_line()




```
### Graphing Coefficients for All Models

```{r}

```
