==> scripts/3_2_figure_visualization.Rmd <==
---
title: "3_2_figure_visualization"
author: "Aspen Coyle"
date: "2024-10-07"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

### Figure Creation

This script creates figures for the final paper.

```{r libraries, message=FALSE, warning=FALSE}
# Add all required libraries here
list.of.packages <- c("tidyverse", "lubridate", "beepr", "maps", "sf", "ggmap", "leaflet", "ggthemes", "cowplot", "easystats", "ggpubr")
# Get names of all required packages that aren't installed
new.packages <- list.of.packages[!(list.of.packages %in% installed.packages()[, "Package"])]
# Install all new packages
if(length(new.packages)) install.packages(new.packages)


# Load all required libraries
lapply(list.of.packages, FUN = function(X) {
  do.call("require", list(X))
})


```

# Figure 1: Map showing survey sites

```{r}
# Read in the pot data to get the locations
loc_dat <- read.csv(file = "../output/ADFG_SE_AK_pot_surveys/cleaned_data/pot_data_with_temperature.csv")

# Remove all except Project, Location, Latitude, and Longitude, rename columns to be easier
loc_dat <- loc_dat %>%
  select("Project", "Location", "Latitude.Decimal.Degrees", "Longitude.Decimal.Degrees") %>%
  dplyr::rename(Latitude = Latitude.Decimal.Degrees, 
         Longitude = Longitude.Decimal.Degrees)

# We've got some weird latitudes and longitudes, let's fix
# Drop all latitudes under 50 and over 60
# Drop all longitudes over -130 and under -150
loc_dat <- loc_dat %>%
  dplyr::filter(Latitude > 50 & Latitude < 60) %>%
  dplyr::filter(Longitude > -150 & Longitude < -130)

# Average lat and long by location
loc_dat <- loc_dat %>%
  group_by(Location, Project) %>%
  summarise(avg_lat = mean(Latitude, na.rm = TRUE),
            avg_long = mean(Longitude, na.rm = TRUE))


# Use data to create a geometry column which has compatible coordinates
map_dat <- loc_dat %>% 
  as.data.frame %>% 
  sf::st_as_sf(coords = c(4, 3))

# Convert this to a file that has the coordinates of each site 
sites_sfc <- st_sfc(map_dat$geometry, crs = 4326)

# Test that it's working and looking good
plot(sites_sfc)

# Register Stadia Maps key
register_stadiamaps("bbdd148e-ddf4-4650-ac62-ec8dfed17f18", write = TRUE)

# Define the boundaries of our basemap
southeast <- get_stadiamap(
  bbox = c(left = -138, bottom = 56, right = -132, top = 59),
  zoom = 8,
  maptype = "stamen_terrain_background")

# Create map of southeast
se_map <- ggmap(southeast) +
  geom_point(data = loc_dat,
             aes(x = avg_long, 
                 y = avg_lat,
                 shape = Project),
             size = 2) +
  xlab ("Longitude") +
  ylab ("Latitude (°N)") +
  guides(shape = FALSE)

# Define boundaries of our Alaska map
alaska <- get_stadiamap(
  bbox = c(left = -180, bottom = 46, right = -120, top = 73),
  zoom = 5,
  maptype = "stamen_terrain_background")

# Create map of Alaska
ak_map <- ggmap(alaska) +
  theme_map() +
  geom_rect(
    xmin = -138,
    ymin = 56, 
    xmax = -132,
    ymax = 59,
    fill = NA,
    colour = "black",
    linewidth = 0.5) +
  theme(legend.position = "none")

# Put AK map on top of SE map, with border around AK map
full_map <- se_map %>%
  ggdraw () +
  draw_plot(ak_map,
            x = 0.141,
            y = 0.079,
            width = 0.45,
            height = 0.45)

# Save plot
ggsave(filename = "../paper/images/Fig1_survey_location_map.png")
```

### Figure 2: Coefficients of General Model

```{r, fig.height=5, fig.width=7}
# Set model filepath
model_filepath <- "../output/ADFG_SE_AK_pot_surveys/models/weighted_models/all_crabs/"

# Read in average model
avg_model <- readRDS(file = paste0(model_filepath, "avg_model.rds"))

# Get basic plot
plot(avg_model)

# Create table of coefficients
coef_tab <- summary(avg_model)$coefmat.full
coef_tab <- as.data.frame(coef_tab)

# Turn rownames into column
coef_tab <- rownames_to_column(coef_tab, var = "variable")

# Remove column header whitespace
colnames(coef_tab) <- str_remove_all(colnames(coef_tab), " ")

# Edit variable column to remove unneeded elements
coef_tab$variable <- coef_tab$variable %>%
  str_replace("^cond\\(", "")
coef_tab$variable <- coef_tab$variable %>%
  str_replace("s.", "")
coef_tab <- dplyr::rename(coef_tab, significant = `Pr(>|z|)`)

# Remove int row
coef_tab <- coef_tab[coef_tab$variable != "(Int))", ]


# Arrange by magnitude of estimated coefficient
coef_tab <- coef_tab %>%
  arrange(desc(abs(Estimate)))

# Change pval to y/n
coef_tab$significant[coef_tab$significant >= 0.05] <- "n"
coef_tab$significant[coef_tab$significant < 0.05] <- "y"

# Create graph. Start with basic image and sorting
coef_tab %>%
  ggplot(aes(x = Estimate, y = reorder(variable, abs(Estimate), decreasing = FALSE), 
             colour = significant)) +
  geom_point() + 
  geom_errorbar(aes(xmin = Estimate - AdjustedSE, 
                    xmax = Estimate + AdjustedSE),
                width = 0.3) +
  geom_vline(xintercept=0, linetype = "dotted") +

  #mess with labels and axes
  scale_y_discrete(labels = c("Latitude", "Leg Condition", "Temperature",
                              "Depth", "Carapace Width", "Sex",
                              "Shell Cond. (Cubic)", "Black Mat", "Shell Cond. (Quadratic)",
                              "Shell Cond. (Linear)")) +
  ylab("Predictor Variable") +
  xlab ("Estimated Coefficient") +

  # mess with theme and appearance
  theme_bw() +
  theme(panel.grid.major = element_blank(),
        panel.grid.minor = element_blank(),
        panel.grid.major.y = element_line(color = "grey",
                                          linewidth = 0.2,
                                          linetype = 2)) +
  theme(plot.title = element_text(hjust = 0.5)) +
  scale_color_manual(values = c("darkgrey", "black")) +
  theme(legend.position = "none")


# Save plot
ggsave(filename = "../paper/images/Fig2_gen_model_coefficients.png")
```

### Figure 3: Graphing General Model Raw Datal

```{r}
# Read in data
crabdat <- read.csv(file = "../output/ADFG_SE_AK_pot_surveys/cleaned_data/crab_data/BCS_examined_crab_with_temperature.csv")


#### Shell Condition -----------------------------------------------------------------
# Calculate means and standard errors
SC_means <- tapply(crabdat$Bitter, crabdat$Shell.Condition, FUN = mean)
SE_vals <- with(crabdat, tapply(Bitter, Shell.Condition, FUN = sd) / sqrt(table(crabdat$Shell.Condition)))

# Create new data table
SC_table <- data.frame(SC_means, SE_vals)
# Rename and reorder columns
SC_table <- SC_table %>%
  dplyr::rename(bitter_mean = SC_means,
         SC = Var1, 
         std_error = Freq) %>%
  relocate(SC)

# Multiply values by 100 to get percentages
SC_table$bitter_mean = SC_table$bitter_mean * 100
SC_table$std_error = SC_table$std_error * 100

# Plot SC
SC_plot <- ggplot(SC_table) +
  geom_bar(aes(x = SC, y = bitter_mean), 
           stat = "identity", 
           fill = "grey")+
  geom_errorbar(aes(x = SC, 
                    ymin = bitter_mean - std_error, 
                    ymax = bitter_mean + std_error),
                width = 0.3) +
  theme_bw() +
  theme(panel.grid.major = element_blank(),
        panel.grid.minor = element_blank()) +
  ylab("Percent of Crabs with BCS") +
  xlab("Shell Condition") +
  scale_x_discrete(labels = c("Light", "New", "Old", "Very Old")) +
  scale_y_continuous(limits = c(0,18))

  
SC_plot
  
#### Black Mat ----------------------------------------

# Calculate means and standard errors
BM_means <- tapply(crabdat$Bitter, crabdat$Blackmat, FUN = mean)
SE_vals <- with(crabdat, tapply(Bitter, Blackmat, FUN = sd) / sqrt(table(crabdat$Blackmat)))

# Create new data table
BM_table <- data.frame(BM_means, SE_vals)
# Rename and reorder columns
BM_table <- BM_table %>%
  dplyr::rename(bitter_mean = BM_means,
         BM = Var1, 
         std_error = Freq) %>%
  relocate(BM)

# Multiply values by 100 to get percentages
BM_table$bitter_mean = BM_table$bitter_mean * 100
BM_table$std_error = BM_table$std_error * 100

# Plot BM
BM_plot <- ggplot(BM_table) +
  geom_bar(aes(x = BM, y = bitter_mean), 
           stat = "identity", 
           fill = "grey")+
  geom_errorbar(aes(x = BM, 
                    ymin = bitter_mean - std_error, 
                    ymax = bitter_mean + std_error),
                width = 0.3) +
  theme_bw() +
  theme(panel.grid.major = element_blank(),
        panel.grid.minor = element_blank()) +
  xlab("Black Mat Status") +
  scale_x_discrete(labels = c("Uninfected", "Infected")) +
  scale_y_continuous(limits = c(0,18)) +
  theme(axis.title.y = element_blank())

  
BM_plot

#### Sex -------------------------------------

# Calculate means and standard errors
sex_means <- tapply(crabdat$Bitter, crabdat$Sex, FUN = mean)
SE_vals <- with(crabdat, tapply(Bitter, Sex, FUN = sd) / sqrt(table(crabdat$Sex)))

# Create new data table
sex_table <- data.frame(sex_means, SE_vals)
# Rename and reorder columns
sex_table <- sex_table %>%
  dplyr::rename(bitter_mean = sex_means,
         sex = Var1, 
         std_error = Freq) %>%
  relocate(sex)

# Multiply values by 100 to get percentages
sex_table$bitter_mean = sex_table$bitter_mean * 100
sex_table$std_error = sex_table$std_error * 100

# Plot sex
sex_plot <- ggplot(sex_table) +
  geom_bar(aes(x = sex, y = bitter_mean), 
           stat = "identity", 
           fill = "grey")+
  geom_errorbar(aes(x = sex, 
                    ymin = bitter_mean - std_error, 
                    ymax = bitter_mean + std_error),
                width = 0.3) +
  theme_bw() +
  theme(panel.grid.major = element_blank(),
        panel.grid.minor = element_blank()) +
  ylab("Percent of Crabs with BCS") +
  xlab("Sex") +
  scale_x_discrete(labels = c("Male", "Female")) +
  scale_y_continuous(limits = c(0,18))

  
sex_plot

#### CW -------------------------------------

# Scale CW separately for each sex
crabdat_f <- crabdat[crabdat$Sex == "2", ]
crabdat_m <- crabdat[crabdat$Sex == "1", ]

crabdat_f$s.CW <- scale(crabdat_f$CW)
crabdat_m$s.CW <- scale(crabdat_m$CW)

crabdat <- rbind(crabdat_f, crabdat_m)

# Get percentile of each scaled CW
crabdat$perc_CW <- ntile(crabdat$s.CW, n = 100)


# Calculate means and standard errors
cw_means <- tapply(crabdat$Bitter, crabdat$perc_CW, FUN = mean)
SE_vals <- with(crabdat, tapply(Bitter, perc_CW, FUN = sd) / sqrt(table(crabdat$perc_CW)))

# Create new data table
cw_table <- data.frame(cw_means, SE_vals)
# Rename and reorder columns
cw_table <- cw_table %>%
  dplyr::rename(bitter_mean = cw_means,
         cw = Var1, 
         std_error = Freq) %>%
  relocate(cw)

# Multiply values by 100 to get percentages
cw_table$bitter_mean = cw_table$bitter_mean * 100
cw_table$std_error = cw_table$std_error * 100

# Convert cw values to numeric
cw_table$cw <- as.numeric(cw_table$cw)

# Plot cw
cw_plot <- ggplot(cw_table) +
  geom_point(aes(x = cw, y = bitter_mean), 
           stat = "identity") +
  geom_errorbar(aes(x = cw, 
                    ymin = bitter_mean - std_error, 
                    ymax = bitter_mean + std_error),
                width = 0.3) +
  theme_bw() +
  theme(panel.grid.major = element_blank(),
        panel.grid.minor = element_blank()) +
  xlab("Sex-Specific Percentile of Carapace Width") +
  scale_y_continuous(limits = c(0,18)) +
  scale_x_continuous(limits = c(0, 100), expand = c(0,1)) +
  theme(axis.title.y = element_blank())


cw_plot


ggdraw() +
  draw_plot(SC_plot, x = 0, y = .5, width = .5, height = .5) +
  draw_plot(BM_plot, x = 0.5, y = .5, width = .5, height = .5) +
  draw_plot(sex_plot, x = 0, y = 0, width = .5, height = .5) +
  draw_plot(cw_plot, x = .5, y = 0, width = .5, height = .5) +
  draw_label(label = "A", x = .1, y = .93) +
  draw_label(label = "B", x = .58, y = .93) +
  draw_label(label = "C", x = .1, y = .43) +
  draw_label(label = "D", x = .58, y = .43) 

ggsave(filename = "../paper/images/Fig3_general_model_data.png")



```


### Figure 4a: Coefficients of Sex-Specific Models

Here, we'll build female-specific and male-specific model coefficient graphs in separate chunks, and then compile them into a single plot

```{r, fig.height=5, fig.width=8}
# Set model filepath
model_filepath <- "../output/ADFG_SE_AK_pot_surveys/models/weighted_models/fem_crabs/"

# Read in average model
avg_model <- readRDS(file = paste0(model_filepath, "avg_model.rds"))

# Get basic plot
plot(avg_model)

# Create table of coefficients
coef_tab <- summary(avg_model)$coefmat.full
coef_tab <- as.data.frame(coef_tab)

# Turn rownames into column
coef_tab <- rownames_to_column(coef_tab, var = "variable")

# Remove column header whitespace
colnames(coef_tab) <- str_remove_all(colnames(coef_tab), " ")

# Edit variable column to remove unneeded elements
coef_tab$variable <- coef_tab$variable %>%
  str_replace("^cond\\(", "")
coef_tab$variable <- coef_tab$variable %>%
  str_replace("s.", "")
coef_tab <- dplyr::rename(coef_tab, significant = `Pr(>|z|)`)

# Remove int row
coef_tab <- coef_tab[coef_tab$variable != "(Int))", ]

# Arrange by magnitude of estimated coefficient
coef_tab <- coef_tab %>%
  arrange(desc(abs(Estimate)))

# Change pval to y/n
coef_tab$significant[coef_tab$significant >= 0.05] <- "n"
coef_tab$significant[coef_tab$significant != "n"] <- "y"

# Create graph. Start with basic image and sorting
fem_mod <- coef_tab %>%
  ggplot(aes(x = Estimate, y = reorder(variable, abs(Estimate), decreasing = FALSE),
             colour = significant)) +
  geom_point() + 
  geom_errorbar(aes(xmin = Estimate - AdjustedSE, 
                    xmax = Estimate + AdjustedSE),
                width = 0.3) +
  geom_vline(xintercept=0, linetype = "dotted") +

  #mess with labels and axes
  scale_y_discrete(labels = c("Latitude", "Leg Condition", "Depth",
                              "Egg Dev. (Uneyed)", "Temperature", "Carapace Width",
                              "Clutch Fullness", "Black Mat", "Egg Dev. (Barren)", 
                              "Egg Dev. (Dead Eggs)", "Egg Dev. (Juvenile)", "Shell Cond. (Cubic)",
                              "Shell Cond. (Quadratic)", "Shell Cond. (Linear)")) +
  scale_x_continuous(limits = c(-5, 1.7)) +
  ylab("Predictor Variable") +
  xlab ("Estimated Coefficient") +
  
  # mess with theme and appearance
  theme_bw() +
  theme(panel.grid.major = element_blank(),
        panel.grid.minor = element_blank(),
        panel.grid.major.y = element_line(color = "grey",
                                          linewidth = 0.2,
                                          linetype = 2)) +
  scale_color_manual(values = c("darkgrey", "black")) +
  theme(legend.position = "none")

# Save plot
ggsave(filename = "../paper/images/Fig4a_female_model_coefficients.png")
```

### Figure 4b: Coefficients of Male Model

```{r, fig.height = 5, fig.width = 8}
# Set model filepath
model_filepath <- "../output/ADFG_SE_AK_pot_surveys/models/weighted_models/male_crabs/"

# Read in average model
avg_model <- readRDS(file = paste0(model_filepath, "avg_model.rds"))

# Get basic plot
plot(avg_model)

# Create table of coefficients
coef_tab <- summary(avg_model)$coefmat.full
coef_tab <- as.data.frame(coef_tab)

# Turn rownames into column
coef_tab <- rownames_to_column(coef_tab, var = "variable")

# Remove column header whitespace
colnames(coef_tab) <- str_remove_all(colnames(coef_tab), " ")

# Edit variable column to remove unneeded elements
coef_tab$variable <- coef_tab$variable %>%
  str_replace("^cond\\(", "")
coef_tab$variable <- coef_tab$variable %>%
  str_replace("s.", "")
coef_tab <- dplyr::rename(coef_tab, significant = `Pr(>|z|)`)

# Remove int row
coef_tab <- coef_tab[coef_tab$variable != "(Int))", ]

# Arrange by magnitude of estimated coefficient
coef_tab <- coef_tab %>%
  arrange(desc(abs(Estimate)))

# Change pval to y/n
coef_tab$significant[coef_tab$significant >= 0.05] <- "n"
coef_tab$significant[coef_tab$significant != "n"] <- "y"

# Create graph. Start with basic image and sorting
male_mod <- coef_tab %>%
  ggplot(aes(x = Estimate, y = reorder(variable, abs(Estimate), decreasing = FALSE),
             colour = significant)) +
  geom_point() + 
  geom_errorbar(aes(xmin = Estimate - AdjustedSE, 
                    xmax = Estimate + AdjustedSE),
                width = 0.3) +
  geom_vline(xintercept=0, linetype = "dotted") +

  #mess with labels and axes
  scale_y_discrete(labels = c("Latitude", "Black Mat", "Temperature",
                              "Depth", "Leg Condition", "Carapace Width",
                              "Shell Cond. (Cubic)", "Maturity", "Shell Cond. (Quadratic)",
                              "Shell Cond. (Linear)")) +
  scale_x_continuous(limits = c(-5, 1.7)) +
  ylab("Predictor Variable") +
  xlab ("Estimated Coefficient") +
  
  # mess with theme and appearance
  theme_bw() +
  theme(panel.grid.major = element_blank(),
        panel.grid.minor = element_blank(),
        panel.grid.major.y = element_line(color = "grey",
                                          linewidth = 0.2,
                                          linetype = 2)) +
  scale_color_manual(values = c("darkgrey", "black")) +
  theme(legend.position = "none")

# Save plot
ggsave(filename = "../paper/images/Fig4b_male_model_coefficients.png")
```

### Figure 4: Create a single figure uniting the female, and male mods

```{r, fig.height=9, fig.width = 10}
ggarrange(fem_mod + rremove("xlab"), 
          male_mod,
          labels = c("F", "M"),
          ncol = 1, nrow = 2,
          align = "h")

# Save plot
ggsave(filename = "../paper/images/Fig_4_sex_specific_models.png")
```


#### Figure 5: Graphing Sex-Specific Raw Data

```{r fig.height=8, fig.width=8}
# Read in data
femdat <- read.csv(file = "../output/ADFG_SE_AK_pot_surveys/cleaned_data/crab_data/female_data_for_model.csv")
mascdat <- read.csv(file = "../output/ADFG_SE_AK_pot_surveys/cleaned_data/crab_data/male_data_for_model.csv")

#### Egg Development -----------------------------------------------------------------
# Calculate means and standard errors
dev_means <- tapply(femdat$Bitter, femdat$Egg.Development, FUN = mean)
SE_vals <- with(femdat, tapply(Bitter, Egg.Development, FUN = sd) / sqrt(table(femdat$Egg.Development)))

# Create new data table
dev_table <- data.frame(dev_means, SE_vals)
# Rename and reorder columns
dev_table <- dev_table %>%
  dplyr::rename(bitter_mean = dev_means,
         dev = Var1, 
         std_error = Freq) %>%
  relocate(dev)

# Multiply values by 100 to get percentages
dev_table$bitter_mean = dev_table$bitter_mean * 100
dev_table$std_error = dev_table$std_error * 100

# Plot dev
dev_plot <- ggplot(dev_table) +
  geom_bar(aes(x = dev, y = bitter_mean), 
           stat = "identity", 
           fill = "grey")+
  geom_errorbar(aes(x = dev, 
                    ymin = bitter_mean - std_error, 
                    ymax = bitter_mean + std_error),
                width = 0.3) +
  theme_bw() +
  theme(panel.grid.major = element_blank(),
        panel.grid.minor = element_blank()) +
  ylab("Percent of Crabs with BCS") +
  xlab("Egg Development") +
  scale_x_discrete(labels = c("Clean", "Matted", "Dead Eggs", "Juvenile", "Uneyed")) +
  scale_y_continuous(limits = c(0,42)) 

dev_plot

#### Clutch Fullness -----------------------------------------------------------------
# Calculate means and standard errors
pct_means <- tapply(femdat$Bitter, femdat$Egg.Percent, FUN = mean)
SE_vals <- with(femdat, tapply(Bitter, Egg.Percent, FUN = sd) / sqrt(table(femdat$Egg.Percent)))

# Create new data table
pct_table <- data.frame(pct_means, SE_vals)
# Rename and reorder columns
pct_table <- pct_table %>%
  dplyr::rename(bitter_mean = pct_means,
         pct = Var1, 
         std_error = Freq) %>%
  relocate(pct)

# Remove values without representation
pct_table$pct <- as.numeric(levels(pct_table$pct))
pct_table <- pct_table %>%
  filter(bitter_mean > 0.01) %>%
  filter(pct != 5)
  
# Multiply values by 100 to get percentages
pct_table$bitter_mean = pct_table$bitter_mean * 100
pct_table$std_error = pct_table$std_error * 100

# Plot pct
pct_plot <- ggplot(pct_table) +
  geom_bar(aes(x = pct, y = bitter_mean), 
           stat = "identity", 
           fill = "grey")+
  geom_errorbar(aes(x = pct, 
                    ymin = bitter_mean - std_error, 
                    ymax = bitter_mean + std_error),
                width = 1.5) +
  theme_bw() +
  theme(panel.grid.major = element_blank(),
        panel.grid.minor = element_blank()) +
  ylab("Percent of Crabs with BCS") +
  xlab("Clutch Fullness (%)") +
  scale_y_continuous(limits = c(0,42)) +
  theme(axis.title.y = element_blank())

pct_plot

# Plot distribution of clutch fullness for infected and uninfected

# Reorder data to display infected and uninfected counts at each CF
CF_distrib <- femdat %>%
  group_by(Egg.Percent) %>%
  summarise(n_rows = length(Bitter),
            n_bitter = sum(Bitter))

# Remove underrepresented cols
# pct_table$pct <- as.numeric(levels(pct_table$pct))
CF_distrib <- CF_distrib %>%
  filter(n_rows > 30)

# Create cols for distrib of bitter and non-bitter
CF_distrib$non_bitter <- CF_distrib$n_rows - CF_distrib$n_bitter
CF_distrib$hist_bitter <- CF_distrib$n_bitter / sum(CF_distrib$n_bitter)
CF_distrib$hist_nb <- CF_distrib$non_bitter / sum(CF_distrib$non_bitter)

# Pivot longer, drop irrelevant columns
CF_distrib <- CF_distrib %>%
  pivot_longer(cols = c(hist_bitter, hist_nb), names_to = "bitter_status", values_to = "hist_vals") %>%
  dplyr::select(c(Egg.Percent, bitter_status, hist_vals))

 # Plot as histogram
CF_hist <- CF_distrib %>%
  ggplot() +
  geom_bar(aes(x = Egg.Percent, y = hist_vals, fill = bitter_status),
           stat = "identity", position = position_dodge()) +
  theme_bw() +
  theme(panel.grid.major = element_blank(),
        panel.grid.minor = element_blank()) +
  xlab("Clutch Fullness (%)") +
  ylab("Frequency") +
  scale_fill_manual("BCS Status", labels = c("Positive", "Negative"), values = c("grey", "black")) +
  theme(legend.position = "inside", legend.position.inside = c(0.2, 0.6)) 
  
CF_hist

#### Maturity (Males) -----------------------------------------------------------------
# Calculate means and standard errors
mat_means <- tapply(mascdat$Bitter, mascdat$mat_stat, FUN = mean)
SE_vals <- with(mascdat, tapply(Bitter, mat_stat, FUN = sd) / sqrt(table(mascdat$mat_stat)))

# Create new data table
mat_table <- data.frame(mat_means, SE_vals)
# Rename and reorder columns
mat_table <- mat_table %>%
  dplyr::rename(bitter_mean = mat_means,
         mat = Var1, 
         std_error = Freq) %>%
  relocate(mat)

# Multiply values by 100 to get percentages
mat_table$bitter_mean = mat_table$bitter_mean * 100
mat_table$std_error = mat_table$std_error * 100

# Plot mat
mat_plot <- ggplot(mat_table) +
  geom_bar(aes(x = mat, y = bitter_mean), 
           stat = "identity", 
           fill = "grey")+
  geom_errorbar(aes(x = mat, 
                    ymin = bitter_mean - std_error, 
                    ymax = bitter_mean + std_error),
                width = 0.3) +
  theme_bw() +
  theme(panel.grid.major = element_blank(),
        panel.grid.minor = element_blank()) +
  ylab("Percent of Crabs with BCS") +
  xlab("Male Maturity") +
  scale_x_discrete(labels = c("Immature", "Mature"))

mat_plot

ggdraw() +
  draw_plot(dev_plot, x = 0, y = 0.5, width = .5, height = .5) +
  draw_plot(pct_plot, x = 0.5, y = 0.5, width = .5, height = .5) +
  draw_plot(mat_plot, x = 0.5, y = 0, width = .5, height = .5) +
  draw_plot(CF_hist, x = 0, y = 0, width = .5, height = .5) +
  draw_label(label = "A", x = .42, y = .95) +
  draw_label(label = "B", x = .92, y = .95) +
  draw_label(label = "C", x = .42, y = .45) +
  draw_label(label = "D", x = .92, y = .45) 
  

# Save plot
ggsave(filename = "../paper/images/Fig_5_sex_specific_models.png")

```
